refactor: split pipeline and runtime responsibilities
