+++ crates/limpid/src/pipeline.rs
@@ -32,2080 +32,28 @@ use crate::event::{BorrowedEvent, OwnedEvent, QueuedEvent};
 use crate::functions::FunctionRegistry;
 use crate::tap::TapRegistry;
 
-// ---------------------------------------------------------------------------
-// Compiled configuration
-// ---------------------------------------------------------------------------
-
-/// A fully resolved configuration ready for execution.
-#[derive(Clone)]
-pub struct CompiledConfig {
-    pub(crate) node_id: Option<String>,
-    pub(crate) node_key: Option<String>,
-    pub inputs: HashMap<String, InputDef>,
-    pub outputs: HashMap<String, OutputDef>,
-    pub processes: HashMap<String, ProcessDef>,
-    pub pipelines: HashMap<String, PipelineDef>,
-    /// User-defined `def function` declarations, indexed by name.
-    /// Registered into the [`FunctionRegistry`] at runtime startup so
-    /// call sites dispatch through the same `(namespace, name)` path
-    /// as built-in primitives.
-    pub functions: HashMap<String, FunctionDef>,
-    pub global_blocks: HashMap<String, Vec<Property>>,
-}
-
-impl CompiledConfig {
-    pub fn from_config(config: Config) -> Result<Self> {
-        let mut inputs = HashMap::new();
-        let mut outputs = HashMap::new();
-        let mut processes = HashMap::new();
-        let mut pipelines = HashMap::new();
-        let mut functions: HashMap<String, FunctionDef> = HashMap::new();
-        let mut global_blocks = HashMap::new();
-
-        for def in config.definitions {
-            match def {
-                Definition::Input(d) => {
-                    if inputs.contains_key(&d.name) {
-                        bail!("duplicate input definition: {}", d.name);
-                    }
-                    inputs.insert(d.name.clone(), d);
-                }
-                Definition::Output(d) => {
-                    if outputs.contains_key(&d.name) {
-                        bail!("duplicate output definition: {}", d.name);
-                    }
-                    outputs.insert(d.name.clone(), d);
-                }
-                Definition::Process(d) => {
-                    if processes.contains_key(&d.name) {
-                        bail!("duplicate process definition: {}", d.name);
-                    }
-                    processes.insert(d.name.clone(), d);
-                }
-                Definition::Pipeline(d) => {
-                    if pipelines.contains_key(&d.name) {
-                        bail!("duplicate pipeline definition: {}", d.name);
-                    }
-                    pipelines.insert(d.name.clone(), d);
-                }
-                Definition::Function(d) => {
-                    if functions.contains_key(&d.name) {
-                        bail!("duplicate function definition: {}", d.name);
-                    }
-                    functions.insert(d.name.clone(), d);
-                }
-            }
-        }
-
-        for block in config.global_blocks {
-            global_blocks.insert(block.name, block.properties);
-        }
-
-        let compiled = Self {
-            node_id: config.node_id,
-            node_key: config.node_key,
-            inputs,
-            outputs,
-            processes,
-            pipelines,
-            functions,
-            global_blocks,
-        };
-        Ok(compiled)
-    }
-
-    /// Validate cross-references: all referenced inputs, outputs, and processes exist.
-    ///
-    /// Takes no `ModuleRegistry` — process names are resolved
-    /// exclusively against user-defined DSL processes (v0.3.0 removed
-    /// the native process layer), and inputs/outputs are validated by
-    /// construction (`ModuleRegistry::create_input`/`create_output`
-    /// already reject unknown types at build time), so there is
-    /// nothing left here that needs the registry.
-    pub fn validate(&self) -> Result<()> {
-        if self.node_key.is_none()
-            && (self
-                .outputs
-                .values()
-                .any(|output| output.properties.type_name() == "ltp")
-                || self
-                    .inputs
-                    .values()
-                    .any(|input| input.properties.type_name() == "ltp"))
-        {
-            bail!("module type 'ltp' requires top-level node_key");
-        }
-        for (name, input) in &self.inputs {
-            if input.properties.type_name() == "ltp" {
-                crate::modules::input::ltp::validate_static_properties(
-                    name,
-                    input.properties.user_properties(),
-                )?;
-            }
-        }
-        crate::modules::input::ltp::validate_listener_groups(
-            self.inputs
-                .iter()
-                .filter(|(_, input)| input.properties.type_name() == "ltp")
-                .map(|(name, input)| (name.as_str(), input.properties.user_properties())),
-        )?;
-        for (name, output) in &self.outputs {
-            if output.properties.type_name() == "ltp" {
-                crate::modules::output::ltp::validate_static_properties(
-                    name,
-                    output.properties.user_properties(),
-                )?;
-            }
-        }
-        for (name, pipeline) in &self.pipelines {
-            for stmt in &pipeline.body {
-                self.validate_pipeline_stmt(name, stmt)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn validate_pipeline_stmt(&self, pipeline_name: &str, stmt: &PipelineStatement) -> Result<()> {
-        match stmt {
-            PipelineStatement::Input(input_names) => {
-                if input_names.is_empty() {
-                    bail!(
-                        "pipeline '{}': input statement has no input names",
-                        pipeline_name
-                    );
-                }
-                let mut seen = std::collections::HashSet::new();
-                for input_name in input_names {
-                    if !self.inputs.contains_key(input_name) {
-                        bail!(
-                            "pipeline '{}': references unknown input '{}'",
-                            pipeline_name,
-                            input_name
-                        );
-                    }
-                    if !seen.insert(input_name.as_str()) {
-                        bail!(
-                            "pipeline '{}': input '{}' listed more than once",
-                            pipeline_name,
-                            input_name
-                        );
-                    }
-                }
-            }
-            PipelineStatement::Output(output_name) => {
-                if !self.outputs.contains_key(output_name) {
-                    bail!(
-                        "pipeline '{}': references unknown output '{}'",
-                        pipeline_name,
-                        output_name
-                    );
-                }
-            }
-            PipelineStatement::ProcessChain(chain) => {
-                for element in chain {
-                    if let ProcessChainElement::Named(proc_name) = element
-                        && !self.processes.contains_key(proc_name)
-                    {
-                        bail!(
-                            "pipeline '{}': references unknown process '{}'. \
-                             Built-in processes were removed in v0.3.0 — use a DSL \
-                             function (e.g. `syslog.parse(ingress)` as a statement) \
-                             or define your own with `def process {{ ... }}`.",
-                            pipeline_name,
-                            proc_name
-                        );
-                    }
-                }
-            }
-            PipelineStatement::If(if_chain) => {
-                for (_, body) in &if_chain.branches {
-                    for item in body {
-                        if let BranchBody::Pipeline(s) = item {
-                            self.validate_pipeline_stmt(pipeline_name, s)?;
-                        }
-                    }
-                }
-                if let Some(else_body) = &if_chain.else_body {
-                    for item in else_body {
-                        if let BranchBody::Pipeline(s) = item {
-                            self.validate_pipeline_stmt(pipeline_name, s)?;
-                        }
-                    }
-                }
-            }
-            PipelineStatement::Switch(_, arms) => {
-                for arm in arms {
-                    for item in &arm.body {
-                        if let BranchBody::Pipeline(s) = item {
-                            self.validate_pipeline_stmt(pipeline_name, s)?;
-                        }
-                    }
-                }
-            }
-            PipelineStatement::Drop | PipelineStatement::Finish | PipelineStatement::Error(_) => {}
-        }
-        Ok(())
-    }
-}
+mod compiled_config;
+pub use compiled_config::CompiledConfig;
 
 // ---------------------------------------------------------------------------
 // Pipeline runner (for --test mode and runtime)
 // ---------------------------------------------------------------------------
 
-/// Trace entry for --test mode output.
-#[derive(Debug)]
-pub struct TraceEntry {
-    pub stage: String,
-    pub label: String,
-    pub detail: String,
-}
+mod execution;
+#[allow(unused_imports)] // Public facade: TraceEntry is consumed by external callers.
+pub use execution::{
+    ErroredEventContext, OutputCapturePolicy, OutputEvent, PipelineRunResult, PipelineTermination,
+    ProcessEvent, TraceEntry, run_pipeline,
+};
+pub(crate) use execution::{run_pipeline_at, run_pipeline_with_process_metrics_at};
 
-/// How a pipeline terminated.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PipelineTermination {
-    /// Reached end of pipeline or explicit `finish`
-    Finished,
-    /// Explicit `drop` statement (event filtered out)
-    Dropped,
-    /// A `process` statement raised a runtime error (unknown identifier,
-    /// type mismatch, regex compile failure, …). The original event is
-    /// surfaced via [`PipelineRunResult::errored`] so the runtime layer
-    /// can route it to the dead-letter queue (operator-configured
-    /// `error_log` JSONL file, or `tracing::error!` fallback). The
-    /// downstream output stream is unaffected — only events that
-    /// finished cleanly reach the configured outputs.
-    Errored,
-}
-
-/// Sum-type failure context surfaced when an event is routed to the
-/// dead-letter queue.
-///
-/// Two flavors distinguish *where* the failure occurred and therefore
-/// what `event` snapshot is meaningful for replay:
-///
-/// - [`Process`](ErroredEventContext::Process) — a `process` step (named
-///   `def process` invocation, inline `process { ... }` block, or a
-///   top-level `error` statement) failed during pipeline execution.
-///   The captured [`ProcessEvent`] holds the original key / ingress / source /
-///   received_at; egress is not snapshotted because at the failure
-///   point it may hold partial output of an earlier process in the
-///   chain, which would confuse replay. Replay re-runs the pipeline
-///   from scratch via `limpidctl inject input <pipeline_input>`.
-///
-/// - [`Output`](ErroredEventContext::Output) — an output sink failed to
-///   accept the event (queue enqueue failure on the runtime side, retry
-///   exhaustion on the sink side, or batched shutdown drain). The
-///   captured [`OutputEvent`] holds both ingress AND egress, because
-///   the pipeline body already finished and the egress is the rendered
-///   payload the sink was about to write. Replay routes through
-///   `limpidctl inject output <output_name>` and the sink's
-///   `consume()` re-runs internal routing.
-///
-/// The Output flavor records *only* the output name — never an
-/// address, destination, path, key, topic, partition, endpoint,
-/// URL, peer, or any other sink-specific routing metadata. Replay
-/// sends the event back through
-/// the named output's `consume()`, which re-routes internally.
-#[derive(Debug, Clone)]
-pub enum ErroredEventContext {
-    Process {
-        /// Wall-clock at which the error was raised.
-        timestamp: chrono::DateTime<chrono::Utc>,
-        /// Pipeline name (from `def pipeline <name>`).
-        pipeline: String,
-        /// Failure site: `<process_name>` for an explicit `def process`
-        /// invocation, `(inline)` for an inline `process { ... }` block,
-        /// `(pipeline)` for an explicit `error` statement at pipeline
-        /// scope, or `(pipeline body)` for a runtime error raised by
-        /// expression evaluation outside any process.
-        site: String,
-        /// Stringified `ProcessError` / `anyhow::Error` from the failure.
-        reason: String,
-        /// Pre-failure event snapshot (key / ingress / source / received_at only).
-        event: ProcessEvent,
-    },
-    Output {
-        /// Wall-clock at which the error was raised.
-        timestamp: chrono::DateTime<chrono::Utc>,
-        /// Pipeline name when known. Runtime-side enqueue failures carry
-        /// the dispatching pipeline; sink-side retry / shutdown records
-        /// leave this blank because they no longer have pipeline context.
-        pipeline: String,
-        /// Failure site: `<output_name>` for retry exhaustion,
-        /// `<output_name> shutdown` for batched shutdown drain, or
-        /// `<output_name> enqueue` for runtime enqueue failure.
-        site: String,
-        /// Stringified failure reason.
-        reason: String,
-        /// Output name — the *only* sink-routing metadata captured.
-        /// Replay = `limpidctl inject output <output_name>`.
-        output_name: String,
-        /// Event snapshot (key + ingress + egress + source + received_at).
-        event: OutputEvent,
-    },
-}
-
-/// Process-flavor event snapshot for the DLQ.
-///
-/// Carries only the fields needed to re-run the pipeline from scratch:
-/// the original ingress bytes, the source socket, and the input
-/// timestamp. Egress is not captured because at a process failure
-/// point it may hold partial output of an earlier process step.
-#[derive(Debug, Clone)]
-pub struct ProcessEvent {
-    key: uuid::Uuid,
-    pub source: std::net::SocketAddr,
-    pub received_at: chrono::DateTime<chrono::Utc>,
-    pub ingress: bytes::Bytes,
-    ltp_stamps: std::sync::Arc<[crate::ltp::HopStamp]>,
-}
-
-/// Output-flavor event snapshot for the DLQ.
-///
-/// Carries both ingress and egress: the pipeline body already finished
-/// and produced an egress payload — replay through `inject output`
-/// hands the egress directly to the sink's `consume()` for
-/// re-rendering / re-shipping.
-#[derive(Debug, Clone)]
-pub struct OutputEvent {
-    key: uuid::Uuid,
-    pub source: std::net::SocketAddr,
-    pub received_at: chrono::DateTime<chrono::Utc>,
-    pub ingress: bytes::Bytes,
-    pub egress: bytes::Bytes,
-    ltp_stamps: std::sync::Arc<[crate::ltp::HopStamp]>,
-}
-
-impl ProcessEvent {
-    /// Snapshot the process-flavor fields from an [`OwnedEvent`].
-    pub fn from_owned(ev: &OwnedEvent) -> Self {
-        Self {
-            key: ev.key(),
-            source: ev.source,
-            received_at: ev.received_at,
-            ingress: ev.ingress.clone(),
-            ltp_stamps: ev.ltp_stamps_arc(),
-        }
-    }
-
-    /// Return the immutable identity of the captured event.
-    pub fn key(&self) -> uuid::Uuid {
-        self.key
-    }
-
-    pub(crate) fn ltp_stamps(&self) -> std::sync::Arc<[crate::ltp::HopStamp]> {
-        std::sync::Arc::clone(&self.ltp_stamps)
-    }
-}
-
-impl OutputEvent {
-    /// Snapshot the output-flavor fields from an [`OwnedEvent`].
-    pub fn from_owned(ev: &OwnedEvent) -> Self {
-        Self {
-            key: ev.key(),
-            source: ev.source,
-            received_at: ev.received_at,
-            ingress: ev.ingress.clone(),
-            egress: ev.egress.clone(),
-            ltp_stamps: ev.ltp_stamps_arc(),
-        }
-    }
-
-    /// Return the immutable identity of the captured event.
-    pub fn key(&self) -> uuid::Uuid {
-        self.key
-    }
-
-    pub(crate) fn ltp_stamps(&self) -> std::sync::Arc<[crate::ltp::HopStamp]> {
-        std::sync::Arc::clone(&self.ltp_stamps)
-    }
-}
-
-impl ErroredEventContext {
-    /// Failure-site accessor.
-    pub fn site(&self) -> &str {
-        match self {
-            Self::Process { site, .. } | Self::Output { site, .. } => site,
-        }
-    }
-
-    /// Reason string accessor.
-    pub fn reason(&self) -> &str {
-        match self {
-            Self::Process { reason, .. } | Self::Output { reason, .. } => reason,
-        }
-    }
-
-    /// Wall-clock timestamp accessor.
-    pub fn timestamp(&self) -> chrono::DateTime<chrono::Utc> {
-        match self {
-            Self::Process { timestamp, .. } | Self::Output { timestamp, .. } => *timestamp,
-        }
-    }
-
-    /// Byte-length hint of the recoverable payload — `egress` on
-    /// Output flavor (that is what a replay would ship), `ingress`
-    /// on Process flavor (that is what a replay would re-enter). Used
-    /// by the `Meta` tracing fallback so operators can correlate
-    /// journald records with their pipeline / output metrics
-    /// without the payload bytes themselves leaving the daemon.
-    pub fn payload_size_hint(&self) -> usize {
-        match self {
-            Self::Process { event, .. } => event.ingress.len(),
-            Self::Output { event, .. } => event.egress.len(),
-        }
-    }
-}
-
-/// Result of running an event through a pipeline.
-///
-/// Execution trace is no longer carried here — see `run_pipeline`'s
-/// `trace: Option<&mut Vec<TraceEntry>>` parameter. `--test-pipeline`
-/// passes its own `Vec` and reads it back directly after the call;
-/// the daemon hot path passes `None` and never allocates one.
-pub struct PipelineRunResult {
-    pub outputs: Vec<(String, QueuedEvent)>,
-    /// True iff at least one `output` statement was reached during
-    /// execution (i.e. `outputs` was non-empty *before* the runtime
-    /// drained it into the per-output queues). Needed because the
-    /// runtime moves `outputs` out of this struct on the way to the
-    /// queue senders, so a later `outputs.is_empty()` check would
-    /// always see `true`. Used to distinguish
-    /// `events_finished` (Finished AND emitted ≥1 output) from
-    /// `events_discarded` (Finished AND emitted nothing).
-    pub had_outputs: bool,
-    pub termination: PipelineTermination,
-    /// DLQ records accumulated during this run. Non-empty iff
-    /// `termination == Errored` from a pipeline-side failure, or when
-    /// the runtime layer appends per-failed-output enqueue records
-    /// (one per failed output). The runtime drains each record into
-    /// the configured `error_log`, or — when none is configured —
-    /// delegates to the `error_log_fallback` ladder helper to emit a
-    /// payload-free operator signal by default (or `Meta` / `Full`
-    /// on explicit opt-in). The tracing line is best-effort, not
-    /// load-bearing recovery.
-    pub errored: Vec<ErroredEventContext>,
-}
-
-/// Pre-compiled metric plan for a pipeline's process invocations.
-/// The plan is validated once at startup, so event execution uses
-/// only O(1) token lookups. Validation rejects shape or identity
-/// mismatches before worker construction; defensive checked access
-/// returns an internal error instead of silently attributing an
-/// invocation to the wrong metric series.
-pub(crate) struct PipelineProcessMetrics {
-    nodes: Vec<ProcessMetricNode>,
-    statements: Vec<PipelineMetricStatement>,
-    #[cfg(test)]
-    selection_trap: std::sync::Arc<MetricNodeSelectionTrapState>,
-}
-
-struct ProcessMetricNode {
-    counters: crate::metrics::ProcessCounters,
-    body_plan: Vec<ProcessMetricStatement>,
-    #[cfg(test)]
-    call_sites: Vec<usize>,
-}
-
-pub(crate) struct RawPipelineProcessMetrics {
-    nodes: Vec<ProcessMetricNode>,
-    identities: Vec<ProcessMetricNodeIdentity>,
-    statements: Vec<PipelineMetricStatement>,
-    #[cfg(test)]
-    selection_trap: std::sync::Arc<MetricNodeSelectionTrapState>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct ProcessMetricNodeIdentity {
-    parent: Option<usize>,
-    step: usize,
-    kind: ProcessMetricNodeKind,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum ProcessMetricNodeKind {
-    Named(String),
-    Inline,
-}
-
-enum PipelineMetricStatement {
-    None,
-    Output(crate::metrics::PipelineOutputTimer),
-    ProcessChain(Vec<usize>),
-    If {
-        branches: Vec<Vec<PipelineMetricStatement>>,
-        else_body: Option<Vec<PipelineMetricStatement>>,
-    },
-    Switch(Vec<Vec<PipelineMetricStatement>>),
-}
-
-impl PipelineProcessMetrics {
-    pub(crate) fn register(
-        pipeline: &PipelineDef,
-        processes: &HashMap<String, ProcessDef>,
-        registry: &crate::metrics::Registry,
-    ) -> anyhow::Result<Self> {
-        Self::compile_raw(pipeline, processes, registry)?.validate(pipeline, processes)
-    }
-
-    pub(crate) fn compile_raw(
-        pipeline: &PipelineDef,
-        processes: &HashMap<String, ProcessDef>,
-        registry: &crate::metrics::Registry,
-    ) -> Result<RawPipelineProcessMetrics, crate::metrics::MetricsError> {
-        let mut builder = ProcessMetricsBuilder {
-            pipeline: &pipeline.name,
-            processes,
-            registry,
-            nodes: Vec::new(),
-            identities: Vec::new(),
-            children: Vec::new(),
-            next_step: 1,
-            output_timers: HashMap::new(),
-        };
-        let statements = builder.pipeline_body(&pipeline.body)?;
-        Ok(RawPipelineProcessMetrics {
-            nodes: builder.nodes,
-            identities: builder.identities,
-            statements,
-            #[cfg(test)]
-            selection_trap: std::sync::Arc::new(MetricNodeSelectionTrapState::default()),
-        })
-    }
-
-    fn select_node(&self, token: usize) -> Option<&ProcessMetricNode> {
-        #[cfg(test)]
-        let armed = self
-            .selection_trap
-            .armed
-            .load(std::sync::atomic::Ordering::Relaxed);
-        #[cfg(test)]
-        if armed {
-            self.selection_trap
-                .total_attempts
-                .fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-        }
-        let node = self.nodes.get(token);
-        #[cfg(test)]
-        if armed && node.is_none() {
-            self.selection_trap
-                .invalid_attempts
-                .fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-        }
-        node
-    }
-
-    #[cfg(test)]
-    pub(crate) fn select_node_for_testing(&self, token: usize) -> bool {
-        self.select_node(token).is_some()
-    }
-}
-
-impl RawPipelineProcessMetrics {
-    pub(crate) fn validate(
-        self,
-        pipeline: &PipelineDef,
-        processes: &HashMap<String, ProcessDef>,
-    ) -> anyhow::Result<PipelineProcessMetrics> {
-        let mut validator = ProcessMetricPlanValidator {
-            raw: &self,
-            processes,
-            next_step: 1,
-            edges: HashMap::new(),
-            validated_nodes: HashSet::new(),
-        };
-        validator.pipeline_body(&pipeline.body, &self.statements)?;
-        if validator.validated_nodes.len() != self.nodes.len() {
-            anyhow::bail!("compiled process metric plan contains unreachable nodes");
-        }
-        Ok(PipelineProcessMetrics {
-            nodes: self.nodes,
-            statements: self.statements,
-            #[cfg(test)]
-            selection_trap: self.selection_trap,
-        })
-    }
-
-    #[cfg(test)]
-    pub(crate) fn root_token_for_testing(&self, step: usize) -> Option<usize> {
-        self.identities
-            .iter()
-            .position(|identity| identity.parent.is_none() && identity.step == step)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn child_token_for_testing(&self, parent: usize, ordinal: usize) -> Option<usize> {
-        self.nodes.get(parent)?.call_sites.get(ordinal).copied()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn metric_node_selection_trap_for_testing(&self) -> MetricNodeSelectionTrap {
-        MetricNodeSelectionTrap(std::sync::Arc::clone(&self.selection_trap))
-    }
-
-    #[cfg(test)]
-    pub(crate) fn remove_root_plan_for_testing(&mut self) {
-        self.statements.clear();
-    }
-
-    #[cfg(test)]
-    pub(crate) fn replace_first_root_plan_with_none_for_testing(&mut self) {
-        if let Some(statement) = self.statements.first_mut() {
-            *statement = PipelineMetricStatement::None;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn invalidate_first_root_token_for_testing(&mut self) {
-        for statement in &mut self.statements {
-            if let PipelineMetricStatement::ProcessChain(tokens) = statement
-                && let Some(token) = tokens.first_mut()
-            {
-                *token = usize::MAX;
-                return;
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn swap_first_two_root_tokens_for_testing(&mut self) {
-        for statement in &mut self.statements {
-            if let PipelineMetricStatement::ProcessChain(tokens) = statement
-                && tokens.len() >= 2
-            {
-                tokens.swap(0, 1);
-                return;
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn invalidate_first_nested_token_for_testing(&mut self) {
-        for node in &mut self.nodes {
-            if let Some(ProcessMetricStatement::Call(token)) = node
-                .body_plan
-                .iter_mut()
-                .find(|statement| matches!(statement, ProcessMetricStatement::Call(_)))
-            {
-                *token = usize::MAX;
-                return;
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn swap_first_two_nested_tokens_for_testing(&mut self) {
-        let positions: Vec<(usize, usize)> = self
-            .nodes
-            .iter()
-            .enumerate()
-            .flat_map(|(node_index, node)| {
-                node.body_plan
-                    .iter()
-                    .enumerate()
-                    .filter(|(_, statement)| matches!(statement, ProcessMetricStatement::Call(_)))
-                    .map(move |(statement_index, _)| (node_index, statement_index))
-            })
-            .take(2)
-            .collect();
-        if let [(left_node, left_statement), (right_node, right_statement)] = positions.as_slice() {
-            let left = match self.nodes[*left_node].body_plan[*left_statement] {
-                ProcessMetricStatement::Call(token) => token,
-                _ => unreachable!(),
-            };
-            let right = match self.nodes[*right_node].body_plan[*right_statement] {
-                ProcessMetricStatement::Call(token) => token,
-                _ => unreachable!(),
-            };
-            self.nodes[*left_node].body_plan[*left_statement] = ProcessMetricStatement::Call(right);
-            self.nodes[*right_node].body_plan[*right_statement] =
-                ProcessMetricStatement::Call(left);
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn invalidate_first_inline_token_for_testing(&mut self) {
-        for statement in &mut self.statements {
-            if let PipelineMetricStatement::ProcessChain(tokens) = statement
-                && let Some(token) = tokens.first_mut()
-            {
-                *token = usize::MAX;
-                return;
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn replace_first_process_body_plan_with_none_for_testing(&mut self) {
-        for node in &mut self.nodes {
-            if let Some(statement) = node
-                .body_plan
-                .iter_mut()
-                .find(|statement| matches!(statement, ProcessMetricStatement::Call(_)))
-            {
-                *statement = ProcessMetricStatement::None;
-                return;
-            }
-        }
-    }
-}
-
-struct ProcessMetricPlanValidator<'a> {
-    raw: &'a RawPipelineProcessMetrics,
-    processes: &'a HashMap<String, ProcessDef>,
-    next_step: usize,
-    edges: HashMap<(usize, String), usize>,
-    validated_nodes: HashSet<usize>,
-}
-
-impl ProcessMetricPlanValidator<'_> {
-    fn pipeline_body(
-        &mut self,
-        statements: &[PipelineStatement],
-        plan: &[PipelineMetricStatement],
-    ) -> anyhow::Result<()> {
-        if statements.len() != plan.len() {
-            anyhow::bail!("compiled pipeline metric plan length mismatch");
-        }
-        for (statement, metric) in statements.iter().zip(plan) {
-            self.pipeline_statement(statement, metric)?;
-        }
-        Ok(())
-    }
-
-    fn pipeline_branch(
-        &mut self,
-        body: &[BranchBody],
-        plan: &[PipelineMetricStatement],
-    ) -> anyhow::Result<()> {
-        if body.len() != plan.len() {
-            anyhow::bail!("compiled pipeline branch metric plan length mismatch");
-        }
-        for (item, metric) in body.iter().zip(plan) {
-            let BranchBody::Pipeline(statement) = item else {
-                anyhow::bail!("process statement found in pipeline metric plan");
-            };
-            self.pipeline_statement(statement, metric)?;
-        }
-        Ok(())
-    }
-
-    fn pipeline_statement(
-        &mut self,
-        statement: &PipelineStatement,
-        metric: &PipelineMetricStatement,
-    ) -> anyhow::Result<()> {
-        match (statement, metric) {
-            (
-                PipelineStatement::ProcessChain(chain),
-                PipelineMetricStatement::ProcessChain(tokens),
-            ) => {
-                if chain.len() != tokens.len() {
-                    anyhow::bail!("compiled root process token count mismatch");
-                }
-                for (site, token) in chain.iter().zip(tokens) {
-                    let step = self.next_step;
-                    self.next_step += 1;
-                    let expected_kind = match site {
-                        ProcessChainElement::Named(name) => {
-                            ProcessMetricNodeKind::Named(name.clone())
-                        }
-                        ProcessChainElement::Inline(_) => ProcessMetricNodeKind::Inline,
-                    };
-                    let identity = self.raw.identities.get(*token).ok_or_else(|| {
-                        anyhow::anyhow!("compiled root process token is out of range")
-                    })?;
-                    if identity.parent.is_some()
-                        || identity.step != step
-                        || identity.kind != expected_kind
-                    {
-                        anyhow::bail!("compiled root process token identity mismatch");
-                    }
-                    match site {
-                        ProcessChainElement::Named(name) => {
-                            let body = self
-                                .processes
-                                .get(name)
-                                .map(|process| process.body.as_slice())
-                                .unwrap_or_default();
-                            self.process_body(*token, body, &mut vec![(name.clone(), *token)])?;
-                        }
-                        ProcessChainElement::Inline(body) => {
-                            self.process_body(*token, body, &mut Vec::new())?;
-                        }
-                    }
-                }
-            }
-            (
-                PipelineStatement::If(chain),
-                PipelineMetricStatement::If {
-                    branches,
-                    else_body,
-                },
-            ) => {
-                if chain.branches.len() != branches.len()
-                    || chain.else_body.is_some() != else_body.is_some()
-                {
-                    anyhow::bail!("compiled pipeline if metric plan shape mismatch");
-                }
-                for ((_, body), branch_plan) in chain.branches.iter().zip(branches) {
-                    self.pipeline_branch(body, branch_plan)?;
-                }
-                if let (Some(body), Some(branch_plan)) =
-                    (chain.else_body.as_deref(), else_body.as_deref())
-                {
-                    self.pipeline_branch(body, branch_plan)?;
-                }
-            }
-            (PipelineStatement::Switch(_, arms), PipelineMetricStatement::Switch(metric_arms)) => {
-                if arms.len() != metric_arms.len() {
-                    anyhow::bail!("compiled pipeline switch metric plan shape mismatch");
-                }
-                for (arm, arm_plan) in arms.iter().zip(metric_arms) {
-                    self.pipeline_branch(&arm.body, arm_plan)?;
-                }
-            }
-            (PipelineStatement::Output(_), PipelineMetricStatement::Output(_)) => {}
-            (
-                PipelineStatement::Input(_)
-                | PipelineStatement::Drop
-                | PipelineStatement::Finish
-                | PipelineStatement::Error(_),
-                PipelineMetricStatement::None,
-            ) => {}
-            _ => anyhow::bail!("compiled pipeline metric statement variant mismatch"),
-        }
-        Ok(())
-    }
-
-    fn process_body(
-        &mut self,
-        parent: usize,
-        statements: &[ProcessStatement],
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> anyhow::Result<()> {
-        if !self.validated_nodes.insert(parent) {
-            return Ok(());
-        }
-        let plan = &self
-            .raw
-            .nodes
-            .get(parent)
-            .ok_or_else(|| anyhow::anyhow!("compiled process node is out of range"))?
-            .body_plan;
-        self.process_statements(parent, statements, plan, ancestors)
-    }
-
-    fn process_statements(
-        &mut self,
-        parent: usize,
-        statements: &[ProcessStatement],
-        plan: &[ProcessMetricStatement],
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> anyhow::Result<()> {
-        if statements.len() != plan.len() {
-            anyhow::bail!("compiled process metric plan length mismatch");
-        }
-        for (statement, metric) in statements.iter().zip(plan) {
-            self.process_statement(parent, statement, metric, ancestors)?;
-        }
-        Ok(())
-    }
-
-    fn process_branch(
-        &mut self,
-        parent: usize,
-        body: &[BranchBody],
-        plan: &[ProcessMetricStatement],
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> anyhow::Result<()> {
-        if body.len() != plan.len() {
-            anyhow::bail!("compiled process branch metric plan length mismatch");
-        }
-        for (item, metric) in body.iter().zip(plan) {
-            let BranchBody::Process(statement) = item else {
-                anyhow::bail!("pipeline statement found in process metric plan");
-            };
-            self.process_statement(parent, statement, metric, ancestors)?;
-        }
-        Ok(())
-    }
-
-    fn process_statement(
-        &mut self,
-        parent: usize,
-        statement: &ProcessStatement,
-        metric: &ProcessMetricStatement,
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> anyhow::Result<()> {
-        match (statement, metric) {
-            (ProcessStatement::ProcessCall(name), ProcessMetricStatement::Call(token)) => {
-                if ancestors.iter().any(|(ancestor, _)| ancestor == name) {
-                    anyhow::bail!("compiled process metric plan contains a recursive process call");
-                }
-                let expected = if let Some(existing) = self.edges.get(&(parent, name.clone())) {
-                    *existing
-                } else {
-                    let step = self.next_step;
-                    self.next_step += 1;
-                    let identity = self.raw.identities.get(*token).ok_or_else(|| {
-                        anyhow::anyhow!("compiled nested process token is out of range")
-                    })?;
-                    if identity.parent != Some(parent)
-                        || identity.step != step
-                        || identity.kind != ProcessMetricNodeKind::Named(name.clone())
-                    {
-                        anyhow::bail!("compiled nested process token identity mismatch");
-                    }
-                    self.edges.insert((parent, name.clone()), *token);
-                    *token
-                };
-                if *token != expected {
-                    anyhow::bail!("compiled process call-site token identity mismatch");
-                }
-                if !self.validated_nodes.contains(token) {
-                    let body = self
-                        .processes
-                        .get(name)
-                        .map(|process| process.body.as_slice())
-                        .unwrap_or_default();
-                    ancestors.push((name.clone(), *token));
-                    self.process_body(*token, body, ancestors)?;
-                    ancestors.pop();
-                }
-            }
-            (
-                ProcessStatement::If(chain),
-                ProcessMetricStatement::If {
-                    branches,
-                    else_body,
-                },
-            ) => {
-                if chain.branches.len() != branches.len()
-                    || chain.else_body.is_some() != else_body.is_some()
-                {
-                    anyhow::bail!("compiled process if metric plan shape mismatch");
-                }
-                for ((_, body), branch_plan) in chain.branches.iter().zip(branches) {
-                    self.process_branch(parent, body, branch_plan, ancestors)?;
-                }
-                if let (Some(body), Some(branch_plan)) =
-                    (chain.else_body.as_deref(), else_body.as_deref())
-                {
-                    self.process_branch(parent, body, branch_plan, ancestors)?;
-                }
-            }
-            (ProcessStatement::Switch(_, arms), ProcessMetricStatement::Switch(metric_arms)) => {
-                if arms.len() != metric_arms.len() {
-                    anyhow::bail!("compiled process switch metric plan shape mismatch");
-                }
-                for (arm, arm_plan) in arms.iter().zip(metric_arms) {
-                    self.process_branch(parent, &arm.body, arm_plan, ancestors)?;
-                }
-            }
-            (
-                ProcessStatement::TryCatch(try_body, catch_body),
-                ProcessMetricStatement::TryCatch {
-                    try_body: try_plan,
-                    catch_body: catch_plan,
-                },
-            ) => {
-                self.process_statements(parent, try_body, try_plan, ancestors)?;
-                self.process_statements(parent, catch_body, catch_plan, ancestors)?;
-            }
-            (
-                ProcessStatement::Assign(_, _)
-                | ProcessStatement::LetBinding(_, _)
-                | ProcessStatement::Drop
-                | ProcessStatement::Error(_)
-                | ProcessStatement::ExprStmt(_),
-                ProcessMetricStatement::None,
-            ) => {}
-            _ => anyhow::bail!("compiled process metric statement variant mismatch"),
-        }
-        Ok(())
-    }
-}
-
+mod metrics_plan;
+use metrics_plan::PipelineMetricStatement;
+pub(crate) use metrics_plan::PipelineProcessMetrics;
 #[cfg(test)]
-#[derive(Default)]
-struct MetricNodeSelectionTrapState {
-    armed: std::sync::atomic::AtomicBool,
-    total_attempts: std::sync::atomic::AtomicUsize,
-    invalid_attempts: std::sync::atomic::AtomicUsize,
-}
-
-#[cfg(test)]
-pub(crate) struct MetricNodeSelectionTrap(std::sync::Arc<MetricNodeSelectionTrapState>);
-
-#[cfg(test)]
-impl MetricNodeSelectionTrap {
-    pub(crate) fn arm_for_testing(&self) {
-        self.0
-            .armed
-            .store(true, std::sync::atomic::Ordering::Relaxed);
-    }
-
-    pub(crate) fn total_token_selections_for_testing(&self) -> usize {
-        self.0
-            .total_attempts
-            .load(std::sync::atomic::Ordering::Relaxed)
-    }
-
-    pub(crate) fn invalid_token_selections_for_testing(&self) -> usize {
-        self.0
-            .invalid_attempts
-            .load(std::sync::atomic::Ordering::Relaxed)
-    }
-}
-
-struct ProcessMetricsBuilder<'a> {
-    pipeline: &'a str,
-    processes: &'a HashMap<String, ProcessDef>,
-    registry: &'a crate::metrics::Registry,
-    nodes: Vec<ProcessMetricNode>,
-    identities: Vec<ProcessMetricNodeIdentity>,
-    children: Vec<HashMap<String, usize>>,
-    next_step: usize,
-    output_timers: HashMap<String, crate::metrics::PipelineOutputTimer>,
-}
-
-impl ProcessMetricsBuilder<'_> {
-    fn pipeline_body(
-        &mut self,
-        body: &[PipelineStatement],
-    ) -> Result<Vec<PipelineMetricStatement>, crate::metrics::MetricsError> {
-        body.iter()
-            .map(|statement| self.pipeline_statement(statement))
-            .collect()
-    }
-
-    fn pipeline_branch(
-        &mut self,
-        body: &[BranchBody],
-    ) -> Result<Vec<PipelineMetricStatement>, crate::metrics::MetricsError> {
-        body.iter()
-            .map(|item| match item {
-                BranchBody::Pipeline(statement) => self.pipeline_statement(statement),
-                BranchBody::Process(_) => Ok(PipelineMetricStatement::None),
-            })
-            .collect()
-    }
-
-    fn pipeline_statement(
-        &mut self,
-        statement: &PipelineStatement,
-    ) -> Result<PipelineMetricStatement, crate::metrics::MetricsError> {
-        match statement {
-            PipelineStatement::Output(name) => {
-                let timer = match self.output_timers.get(name) {
-                    Some(timer) => timer.clone(),
-                    None => {
-                        let timer = crate::metrics::PipelineOutputTimer::register(
-                            self.registry,
-                            self.pipeline,
-                            name,
-                        )?;
-                        self.output_timers.insert(name.clone(), timer.clone());
-                        timer
-                    }
-                };
-                Ok(PipelineMetricStatement::Output(timer))
-            }
-            PipelineStatement::ProcessChain(chain) => {
-                let mut nodes = Vec::with_capacity(chain.len());
-                for element in chain {
-                    let (name, kind, body) = match element {
-                        ProcessChainElement::Named(name) => (
-                            name.as_str(),
-                            ProcessMetricNodeKind::Named(name.clone()),
-                            self.processes
-                                .get(name)
-                                .map(|process| process.body.as_slice()),
-                        ),
-                        ProcessChainElement::Inline(body) => (
-                            "(inline)",
-                            ProcessMetricNodeKind::Inline,
-                            Some(body.as_slice()),
-                        ),
-                    };
-                    nodes.push(self.add_node(
-                        None,
-                        kind,
-                        name,
-                        format!("/{name}"),
-                        body,
-                        &mut Vec::new(),
-                    )?);
-                }
-                Ok(PipelineMetricStatement::ProcessChain(nodes))
-            }
-            PipelineStatement::If(chain) => {
-                let branches = chain
-                    .branches
-                    .iter()
-                    .map(|(_, body)| self.pipeline_branch(body))
-                    .collect::<Result<Vec<_>, _>>()?;
-                let else_body = chain
-                    .else_body
-                    .as_deref()
-                    .map(|body| self.pipeline_branch(body))
-                    .transpose()?;
-                Ok(PipelineMetricStatement::If {
-                    branches,
-                    else_body,
-                })
-            }
-            PipelineStatement::Switch(_, arms) => Ok(PipelineMetricStatement::Switch(
-                arms.iter()
-                    .map(|arm| self.pipeline_branch(&arm.body))
-                    .collect::<Result<Vec<_>, _>>()?,
-            )),
-            _ => Ok(PipelineMetricStatement::None),
-        }
-    }
-
-    fn add_node(
-        &mut self,
-        parent: Option<usize>,
-        kind: ProcessMetricNodeKind,
-        name: &str,
-        path: String,
-        body: Option<&[ProcessStatement]>,
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> Result<usize, crate::metrics::MetricsError> {
-        let step = self.next_step;
-        self.next_step += 1;
-        let id = self.nodes.len();
-        self.nodes.push(ProcessMetricNode {
-            counters: crate::metrics::ProcessCounters::register(
-                self.registry,
-                self.pipeline,
-                step,
-                &path,
-                name,
-            )?,
-            body_plan: Vec::new(),
-            #[cfg(test)]
-            call_sites: Vec::new(),
-        });
-        self.identities
-            .push(ProcessMetricNodeIdentity { parent, step, kind });
-        self.children.push(HashMap::new());
-        ancestors.push((name.to_owned(), id));
-        if let Some(body) = body {
-            self.nodes[id].body_plan = self.process_body(id, &path, body, ancestors)?;
-        }
-        ancestors.pop();
-        Ok(id)
-    }
-
-    fn process_body(
-        &mut self,
-        parent: usize,
-        parent_path: &str,
-        body: &[ProcessStatement],
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> Result<Vec<ProcessMetricStatement>, crate::metrics::MetricsError> {
-        body.iter()
-            .map(|statement| self.process_statement(parent, parent_path, statement, ancestors))
-            .collect()
-    }
-
-    fn process_statement(
-        &mut self,
-        parent: usize,
-        parent_path: &str,
-        statement: &ProcessStatement,
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> Result<ProcessMetricStatement, crate::metrics::MetricsError> {
-        match statement {
-            ProcessStatement::ProcessCall(name) => {
-                let child = if let Some(child) = self.children[parent].get(name) {
-                    *child
-                } else if let Some((position, _)) = ancestors
-                    .iter()
-                    .enumerate()
-                    .find(|(_, (ancestor, _))| ancestor == name)
-                {
-                    let mut path: Vec<String> = ancestors[position..]
-                        .iter()
-                        .map(|(ancestor, _)| ancestor.clone())
-                        .collect();
-                    path.push(name.clone());
-                    return Err(crate::metrics::MetricsError::ProcessCallCycle { path });
-                } else {
-                    let body = self
-                        .processes
-                        .get(name)
-                        .map(|process| process.body.as_slice());
-                    let child = self.add_node(
-                        Some(parent),
-                        ProcessMetricNodeKind::Named(name.clone()),
-                        name,
-                        format!("{parent_path}/{name}"),
-                        body,
-                        ancestors,
-                    )?;
-                    self.children[parent].insert(name.clone(), child);
-                    child
-                };
-                #[cfg(test)]
-                self.nodes[parent].call_sites.push(child);
-                Ok(ProcessMetricStatement::Call(child))
-            }
-            ProcessStatement::If(chain) => {
-                let branches = chain
-                    .branches
-                    .iter()
-                    .map(|(_, body)| self.process_branch(parent, parent_path, body, ancestors))
-                    .collect::<Result<Vec<_>, _>>()?;
-                let else_body = chain
-                    .else_body
-                    .as_deref()
-                    .map(|body| self.process_branch(parent, parent_path, body, ancestors))
-                    .transpose()?;
-                Ok(ProcessMetricStatement::If {
-                    branches,
-                    else_body,
-                })
-            }
-            ProcessStatement::Switch(_, arms) => Ok(ProcessMetricStatement::Switch(
-                arms.iter()
-                    .map(|arm| self.process_branch(parent, parent_path, &arm.body, ancestors))
-                    .collect::<Result<Vec<_>, _>>()?,
-            )),
-            ProcessStatement::TryCatch(try_body, catch_body) => {
-                Ok(ProcessMetricStatement::TryCatch {
-                    try_body: self.process_body(parent, parent_path, try_body, ancestors)?,
-                    catch_body: self.process_body(parent, parent_path, catch_body, ancestors)?,
-                })
-            }
-            _ => Ok(ProcessMetricStatement::None),
-        }
-    }
-
-    fn process_branch(
-        &mut self,
-        parent: usize,
-        parent_path: &str,
-        body: &[BranchBody],
-        ancestors: &mut Vec<(String, usize)>,
-    ) -> Result<Vec<ProcessMetricStatement>, crate::metrics::MetricsError> {
-        body.iter()
-            .map(|item| match item {
-                BranchBody::Process(statement) => {
-                    self.process_statement(parent, parent_path, statement, ancestors)
-                }
-                BranchBody::Pipeline(_) => Ok(ProcessMetricStatement::None),
-            })
-            .collect()
-    }
-}
+pub(crate) use metrics_plan::{MetricNodeSelectionTrap, RawPipelineProcessMetrics};
 
 /// A process registry backed by compiled DSL process definitions.
-///
-/// Only user-defined `def process { ... }` blocks resolve here.
-/// Built-in processes were removed in v0.3.0 — former native
-/// transforms are now DSL functions (`syslog.parse`, `parse_json`,
-/// `regex_replace`, …) invoked via expression statements.
-struct DslProcessRegistry<'a> {
-    processes: &'a HashMap<String, ProcessDef>,
-    funcs: &'a FunctionRegistry,
-    tap: Option<&'a TapRegistry>,
-    process_metrics: Option<&'a PipelineProcessMetrics>,
-}
-
-impl<'a> DslProcessRegistry<'a> {
-    fn new(
-        processes: &'a HashMap<String, ProcessDef>,
-        funcs: &'a FunctionRegistry,
-        tap: Option<&'a TapRegistry>,
-        process_metrics: Option<&'a PipelineProcessMetrics>,
-    ) -> Self {
-        Self {
-            processes,
-            funcs,
-            tap,
-            process_metrics,
-        }
-    }
-
-    fn call_node<'bump>(
-        &self,
-        metric_token: Option<usize>,
-        name: &str,
-        event: BorrowedEvent<'bump>,
-        arena: &'bump EventArena<'bump>,
-    ) -> std::result::Result<Option<BorrowedEvent<'bump>>, ProcessError> {
-        let metric_node = match (self.process_metrics, metric_token) {
-            (Some(metrics), Some(token)) => Some(metrics.select_node(token).ok_or_else(|| {
-                ProcessError::Failed("compiled process metric token is out of range".to_owned())
-            })?),
-            (Some(_), None) => {
-                return Err(ProcessError::Failed(
-                    "compiled process metric token is missing".to_owned(),
-                ));
-            }
-            (None, _) => None,
-        };
-        if let Some(node) = metric_node {
-            node.counters.start();
-        }
-
-        let result = if let Some(process_def) = self.processes.get(name) {
-            trace!("process '{}' (user-defined): executing", name);
-            match metric_node {
-                Some(node) => exec_process_body_with_metric_plan(
-                    &process_def.body,
-                    &node.body_plan,
-                    event,
-                    self,
-                    self.funcs,
-                    arena,
-                ),
-                None => exec_process_body(&process_def.body, event, self, self.funcs, arena),
-            }
-            .map_err(|error| ProcessError::Failed(error.to_string()))
-        } else {
-            tracing::warn!(
-                "unknown process '{}', passing event through unchanged",
-                name
-            );
-            Ok(ExecResult::Continue(event))
-        };
-
-        match result {
-            Ok(ExecResult::Continue(event)) => {
-                if let Some(node) = metric_node {
-                    node.counters.continued();
-                }
-                trace!("process '{}': ok", name);
-                self.emit_tap(name, &event);
-                Ok(Some(event))
-            }
-            Ok(ExecResult::Dropped) => {
-                if let Some(node) = metric_node {
-                    node.counters.dropped();
-                }
-                trace!("process '{}': dropped", name);
-                Ok(None)
-            }
-            Err(error) => {
-                if let Some(node) = metric_node {
-                    node.counters.errored();
-                }
-                Err(error)
-            }
-        }
-    }
-}
-
-impl ProcessRegistry for DslProcessRegistry<'_> {
-    fn call<'bump>(
-        &self,
-        name: &str,
-        event: BorrowedEvent<'bump>,
-        arena: &'bump EventArena<'bump>,
-    ) -> std::result::Result<Option<BorrowedEvent<'bump>>, ProcessError> {
-        self.call_node(None, name, event, arena)
-    }
-}
-
-impl CompiledProcessRegistry for DslProcessRegistry<'_> {
-    fn call_pre_resolved<'bump>(
-        &self,
-        name: &str,
-        metric_token: usize,
-        event: BorrowedEvent<'bump>,
-        arena: &'bump EventArena<'bump>,
-    ) -> std::result::Result<Option<BorrowedEvent<'bump>>, ProcessError> {
-        self.call_node(Some(metric_token), name, event, arena)
-    }
-}
-
-impl DslProcessRegistry<'_> {
-    fn emit_tap<'bump>(&self, process_name: &str, event: &BorrowedEvent<'bump>) {
-        if let Some(tap) = self.tap {
-            let key = format!("process {}", process_name);
-            // Avoid the per-event `to_owned()` workspace clone unless a
-            // tap subscriber is actually attached. `is_subscribed`
-            // collapses to a single relaxed atomic load on the hot path
-            // (no lock when the registry isn't being mutated).
-            if tap.is_subscribed(&key) {
-                let owned = event.to_owned();
-                tap.try_emit(&key, &owned);
-            }
-        }
-    }
-}
-
-/// Run a single event through a pipeline definition.
-///
-/// After this change the executor never resolves an output sink — the `output`
-/// statement just enqueues the owned event, and render happens
-/// consumer-side inside each sink's `Output::consume`. The previous
-/// `output_sinks: &HashMap<String, Arc<dyn Output>>` parameter is gone
-/// as part of that cleanup.
-///
-/// `trace` collects a human-readable execution trace for
-/// `--test-pipeline` (see `main.rs::run_test`). Pass `None` on the
-/// daemon hot path (`runtime.rs::run_pipeline_with_outputs`) — every
-/// trace push site (and the `format!`/`to_string` calls that build
-/// its fields) is gated on `trace.is_some()` so no throwaway
-/// formatting work runs when nothing will read it. The collected
-/// entries live in the caller's `Vec`; `PipelineRunResult` does not
-/// carry them.
-#[allow(clippy::too_many_arguments)]
-pub fn run_pipeline(
-    pipeline: &PipelineDef,
-    event: &OwnedEvent,
-    config: &CompiledConfig,
-    funcs: &FunctionRegistry,
-    tap: Option<&TapRegistry>,
-    trace: Option<&mut Vec<TraceEntry>>,
-    output_capture: OutputCapturePolicy<'_>,
-    bump: &mut bumpalo::Bump,
-) -> Result<PipelineRunResult> {
-    run_pipeline_at(
-        pipeline,
-        event,
-        config,
-        funcs,
-        tap,
-        trace,
-        output_capture,
-        bump,
-        crate::time::UnixNanos::now(),
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn run_pipeline_at(
-    pipeline: &PipelineDef,
-    event: &OwnedEvent,
-    config: &CompiledConfig,
-    funcs: &FunctionRegistry,
-    tap: Option<&TapRegistry>,
-    trace: Option<&mut Vec<TraceEntry>>,
-    output_capture: OutputCapturePolicy<'_>,
-    bump: &mut bumpalo::Bump,
-    dispatch_started_at: crate::time::UnixNanos,
-) -> Result<PipelineRunResult> {
-    run_pipeline_inner(
-        pipeline,
-        event,
-        config,
-        funcs,
-        tap,
-        trace,
-        output_capture,
-        bump,
-        None,
-        dispatch_started_at,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn run_pipeline_with_process_metrics_at(
-    pipeline: &PipelineDef,
-    event: &OwnedEvent,
-    config: &CompiledConfig,
-    funcs: &FunctionRegistry,
-    tap: Option<&TapRegistry>,
-    trace: Option<&mut Vec<TraceEntry>>,
-    output_capture: OutputCapturePolicy<'_>,
-    bump: &mut bumpalo::Bump,
-    process_metrics: &PipelineProcessMetrics,
-    dispatch_started_at: crate::time::UnixNanos,
-) -> Result<PipelineRunResult> {
-    run_pipeline_inner(
-        pipeline,
-        event,
-        config,
-        funcs,
-        tap,
-        trace,
-        output_capture,
-        bump,
-        Some(process_metrics),
-        dispatch_started_at,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn run_pipeline_inner(
-    pipeline: &PipelineDef,
-    event: &OwnedEvent,
-    config: &CompiledConfig,
-    funcs: &FunctionRegistry,
-    tap: Option<&TapRegistry>,
-    trace: Option<&mut Vec<TraceEntry>>,
-    output_capture: OutputCapturePolicy<'_>,
-    bump: &mut bumpalo::Bump,
-    process_metrics: Option<&PipelineProcessMetrics>,
-    dispatch_started_at: crate::time::UnixNanos,
-) -> Result<PipelineRunResult> {
-    let registry = DslProcessRegistry::new(&config.processes, funcs, tap, process_metrics);
-    let mut trace = trace;
-    let mut outputs = Vec::new();
-
-    // Log initial state — formatted from `event` while it's still in
-    // owned form, before we view it into the arena. The `format!` /
-    // `from_utf8_lossy` here only run when a caller actually wants a
-    // trace (--test-pipeline); the daemon hot path passes `None` and
-    // skips this entirely.
-    if let Some(trace) = trace.as_mut() {
-        trace.push(TraceEntry {
-            stage: "input".into(),
-            label: String::new(),
-            detail: format!("ingress: {}", String::from_utf8_lossy(&event.ingress)),
-        });
-    }
-
-    // Per-event arena. The entire `Value` tree built during execution
-    // (HashLits, parser outputs, workspace mutations) lives in `bump`
-    // and is reset to offset zero by the caller after this function
-    // returns — see `runtime::run_pipeline_workers`. The `Bump` itself
-    // is owned by the per-input pipeline-worker task and reused
-    // across events, so the underlying chunk-group is malloc'd once
-    // at task startup and never again on the hot path. This
-    // eliminates the xzm-zone-lock contention that capped
-    // multi-pipeline scaling at ~2.4× / 4 cores on v0.6.0 (where
-    // every event called `Bump::new` and the system allocator
-    // serialised concurrent malloc/free across pipelines).
-    let arena = EventArena::new(bump);
-    let bevent = event.view_in(&arena);
-
-    let mut errored: Vec<ErroredEventContext> = Vec::new();
-    let exec_ctx = PipelineExecCtx {
-        pipeline_name: &pipeline.name,
-        registry: &registry,
-        funcs,
-        arena: &arena,
-        output_capture,
-        dispatch_started_at,
-    };
-    let mut exec_out = PipelineExecOut {
-        trace,
-        outputs: &mut outputs,
-        errored: &mut errored,
-    };
-    let (_, termination) = exec_pipeline_body(
-        &pipeline.body,
-        process_metrics.map(|metrics| metrics.statements.as_slice()),
-        bevent,
-        &exec_ctx,
-        &mut exec_out,
-    )?;
-
-    let had_outputs = !outputs.is_empty();
-    Ok(PipelineRunResult {
-        outputs,
-        had_outputs,
-        termination,
-        errored,
-    })
-}
-
-/// Whether each per-output `OwnedEvent` snapshot pushed at
-/// `output` statements carries the pipeline's per-event `workspace`
-/// or drops it on the floor.
-///
-/// The snapshot is what the downstream queue transports. A memory
-/// queue's consumer only reads `egress` (with `file` also reading
-/// `source` / `received_at` and `kafka` optionally `source.ip`),
-/// and every DLQ record projection stores only `OutputEvent`'s four
-/// fields, so the workspace deep-clone was pure overhead on the
-/// hot path — the largest contributor to the D-shape throughput
-/// regression the `b7625bb` refactor introduced (see the release
-/// notes for 0.7.10). Two consumers of the snapshot do still need
-/// the workspace, and this policy names them explicitly rather
-/// than inferring from ambient state:
-///
-/// - **Disk-backed queues** serialise the full `Event` JSON to the
-///   WAL and rehydrate the workspace on replay. Skipping the
-///   capture there would silently change on-disk semantics on the
-///   next restart.
-/// - **`--test-pipeline`** shows the resulting `OwnedEvent`
-///   (workspace included) in its CLI output so operators can see
-///   what a pipeline produced at each sink boundary.
-///
-/// The `tap output --json` path is *not* a workspace consumer under
-/// 0.7.10's tap contract: the tap projection strips `workspace`
-/// unconditionally on the emit side, so both memory and disk
-/// queues expose the same tap-JSON shape regardless of what the
-/// snapshot carries.
-#[derive(Debug, Clone, Copy)]
-pub enum OutputCapturePolicy<'a> {
-    /// Strip the workspace from every output snapshot. Used by
-    /// unit tests that don't care about workspace round-trip.
-    #[allow(dead_code)]
-    StripAll,
-    /// Capture the workspace on every output snapshot. Used by
-    /// `--test-pipeline` so the CLI display shows what the
-    /// pipeline actually built.
-    CaptureAll,
-    /// Capture the workspace only on the named outputs (those whose
-    /// queue is disk-backed). The daemon hot path builds this set
-    /// once at startup from the compiled config's queue kinds and
-    /// hands it in per event.
-    DiskOnly(&'a std::collections::HashSet<String>),
-}
-
-impl<'a> OutputCapturePolicy<'a> {
-    fn should_capture_workspace(&self, output_name: &str) -> bool {
-        match self {
-            Self::StripAll => false,
-            Self::CaptureAll => true,
-            Self::DiskOnly(disk) => disk.contains(output_name),
-        }
-    }
-}
-
-/// Immutable shared context threaded through the pipeline executor.
-///
-/// `pipeline_name` is here purely so a process-runtime error can
-/// populate the [`ErroredEventContext`] surfaced in [`PipelineExecOut::errored`].
-///
-/// `arena` is the per-event bump arena — the same one
-/// `run_pipeline` opened on the stack. The reference itself is held at
-/// `'bump` so closures and primitive impls allocating into it can
-/// produce values that live for the rest of the pipeline body.
-///
-/// `output_capture` decides per output whether the snapshot pushed
-/// to `PipelineExecOut::outputs` carries the workspace — see
-/// [`OutputCapturePolicy`].
-struct PipelineExecCtx<'a, 'bump: 'a> {
-    pipeline_name: &'a str,
-    registry: &'a DslProcessRegistry<'a>,
-    funcs: &'a FunctionRegistry,
-    arena: &'bump EventArena<'bump>,
-    output_capture: OutputCapturePolicy<'a>,
-    dispatch_started_at: crate::time::UnixNanos,
-}
-
-/// Mutable accumulators threaded through the pipeline executor:
-/// trace entries, output queue pushes, and the optional errored event
-/// context. Bundled together to keep the recursive helpers under
-/// clippy's `too_many_arguments` threshold and to make the executor's
-/// "what comes out" surface explicit.
-///
-/// Outputs and errored contexts are heap-owned — they cross the
-/// per-event arena boundary on the way out of `run_pipeline`.
-struct PipelineExecOut<'a> {
-    /// `Some` only on the `--test-pipeline` path (see `run_pipeline`'s
-    /// `trace` parameter). `None` on the daemon hot path, where no
-    /// caller ever reads `PipelineRunResult::trace` — keeping this an
-    /// `Option` (rather than always allocating a `Vec`) lets every
-    /// push site below skip its `format!` / `to_string` formatting
-    /// work entirely when tracing isn't requested, instead of
-    /// computing throwaway `String`s on every event just to push them
-    /// into a `Vec` nobody drains.
-    trace: Option<&'a mut Vec<TraceEntry>>,
-    outputs: &'a mut Vec<(String, QueuedEvent)>,
-    errored: &'a mut Vec<ErroredEventContext>,
-}
-
-impl PipelineExecOut<'_> {
-    /// Push a trace entry, built lazily from `f`, iff tracing is
-    /// enabled. `f` is only invoked when `trace` is `Some`, so the
-    /// `format!`/`to_string` work that builds `TraceEntry` fields
-    /// never runs on the daemon hot path.
-    fn push_trace(&mut self, f: impl FnOnce() -> TraceEntry) {
-        if let Some(trace) = self.trace.as_deref_mut() {
-            trace.push(f());
-        }
-    }
-}
-
-/// Execute a pipeline body (sequence of pipeline statements).
-/// Returns (remaining event if any, how the pipeline terminated).
-fn exec_pipeline_body<'bump>(
-    stmts: &[PipelineStatement],
-    metric_stmts: Option<&[PipelineMetricStatement]>,
-    mut event: BorrowedEvent<'bump>,
-    ctx: &PipelineExecCtx<'_, 'bump>,
-    out: &mut PipelineExecOut<'_>,
-) -> Result<(Option<BorrowedEvent<'bump>>, PipelineTermination)> {
-    if ctx.registry.process_metrics.is_some() {
-        let metric_stmts = metric_stmts
-            .ok_or_else(|| anyhow::anyhow!("compiled pipeline metric plan is missing"))?;
-        if metric_stmts.len() != stmts.len() {
-            bail!("compiled pipeline metric plan length does not match the pipeline body");
-        }
-    }
-    for (index, stmt) in stmts.iter().enumerate() {
-        let metric_stmt = match ctx.registry.process_metrics {
-            Some(_) => Some(
-                metric_stmts
-                    .and_then(|metrics| metrics.get(index))
-                    .ok_or_else(|| anyhow::anyhow!("compiled pipeline metric entry is missing"))?,
-            ),
-            None => None,
-        };
-        match exec_pipeline_stmt(stmt, metric_stmt, event, ctx, out)? {
-            (Some(e), _) => event = e,
-            (None, term) => return Ok((None, term)),
-        }
-    }
-    Ok((Some(event), PipelineTermination::Finished))
-}
-
-fn exec_pipeline_stmt<'bump>(
-    stmt: &PipelineStatement,
-    metric_stmt: Option<&PipelineMetricStatement>,
-    event: BorrowedEvent<'bump>,
-    ctx: &PipelineExecCtx<'_, 'bump>,
-    out: &mut PipelineExecOut<'_>,
-) -> Result<(Option<BorrowedEvent<'bump>>, PipelineTermination)> {
-    let cont = |event| Ok((Some(event), PipelineTermination::Finished));
-    let dropped = || Ok((None, PipelineTermination::Dropped));
-    let finished = || Ok((None, PipelineTermination::Finished));
-
-    match stmt {
-        PipelineStatement::Input(_) => cont(event),
-
-        PipelineStatement::Error(msg_expr) => {
-            // Render the optional message and route the event to the
-            // error_log via PipelineTermination::Errored, mirroring how
-            // a process-level Err lands in the DLQ.
-            let msg = match msg_expr {
-                Some(e) => value_to_string(&eval_expr(e, &event, ctx.funcs, ctx.arena)?),
-                None => "explicit error routing".to_string(),
-            };
-            tracing::warn!(
-                "pipeline '{}': error '{}' — event routed to error_log",
-                ctx.pipeline_name,
-                msg
-            );
-            out.push_trace(|| TraceEntry {
-                stage: "error".into(),
-                label: msg.clone(),
-                detail: "event → error_log".into(),
-            });
-            // Cross to owned form for the DLQ context (which must
-            // outlive the per-event arena).
-            let owned = event.to_owned();
-            out.errored.push(ErroredEventContext::Process {
-                timestamp: chrono::Utc::now(),
-                pipeline: ctx.pipeline_name.to_string(),
-                site: "(pipeline)".to_string(),
-                reason: msg,
-                event: ProcessEvent::from_owned(&owned),
-            });
-            Ok((None, PipelineTermination::Errored))
-        }
-
-        PipelineStatement::ProcessChain(chain) => {
-            let mut current = event;
-            let metric_nodes = match metric_stmt {
-                Some(PipelineMetricStatement::ProcessChain(nodes)) => Some(nodes.as_slice()),
-                _ => None,
-            };
-            for (index, element) in chain.iter().enumerate() {
-                let metric_token = metric_nodes.and_then(|nodes| nodes.get(index)).copied();
-                match element {
-                    ProcessChainElement::Named(name) => {
-                        // Snapshot the pre-call view before the registry
-                        // consumes the borrowed event — the Err arm
-                        // needs a stable, DLQ-ready event. Use an
-                        // arena-local shallow snapshot rather than
-                        // `to_owned`: the success path (dominant on the
-                        // hot path) only needs the snapshot to survive
-                        // until the `match` returns `Ok(...)`, and
-                        // paying the heap materialization of the
-                        // workspace `HashMap` + `Value` tree on every
-                        // successful process call was a significant
-                        // fraction of the runtime for multi-process
-                        // pipelines where `parse | enrich | route`-shaped
-                        // chains re-entered process #2+ with a populated
-                        // workspace. `snapshot_in` bumps `Bytes`
-                        // refcounts and copies the workspace index vec;
-                        // the deep `to_owned` clone happens only in the
-                        // Err arm below, where the DLQ record's owned
-                        // event has to cross the arena boundary anyway.
-                        let backup_view = current.snapshot_in(ctx.arena);
-                        let call_result = match ctx.registry.process_metrics {
-                            Some(_) => {
-                                let token = metric_token.ok_or_else(|| {
-                                    anyhow::anyhow!("compiled root process token is missing")
-                                })?;
-                                ctx.registry
-                                    .call_pre_resolved(name, token, current, ctx.arena)
-                            }
-                            None => ctx.registry.call(name, current, ctx.arena),
-                        };
-                        match call_result {
-                            Ok(Some(e)) => {
-                                out.push_trace(|| TraceEntry {
-                                    stage: "process".into(),
-                                    label: name.clone(),
-                                    detail: "ok".into(),
-                                });
-                                current = e;
-                            }
-                            Ok(None) => {
-                                out.push_trace(|| TraceEntry {
-                                    stage: "process".into(),
-                                    label: name.clone(),
-                                    detail: "dropped".into(),
-                                });
-                                return dropped();
-                            }
-                            Err(e) => {
-                                tracing::warn!(
-                                    "process '{}': {} — event routed to error_log",
-                                    name,
-                                    e
-                                );
-                                out.push_trace(|| TraceEntry {
-                                    stage: "process".into(),
-                                    label: name.clone(),
-                                    detail: format!("error: {} (event → error_log)", e),
-                                });
-                                out.errored.push(ErroredEventContext::Process {
-                                    timestamp: chrono::Utc::now(),
-                                    pipeline: ctx.pipeline_name.to_string(),
-                                    site: name.clone(),
-                                    reason: e.to_string(),
-                                    // Cross the arena boundary here on
-                                    // the (rare) failure path only: the
-                                    // owned DLQ record outlives the
-                                    // per-event arena.
-                                    event: ProcessEvent::from_owned(&backup_view.to_owned()),
-                                });
-                                return Ok((None, PipelineTermination::Errored));
-                            }
-                        }
-                    }
-                    ProcessChainElement::Inline(body) => {
-                        // Same rationale as the Named arm: arena-local
-                        // shallow snapshot for the DLQ Err path; the
-                        // heap materialization happens only on failure.
-                        let backup_view = current.snapshot_in(ctx.arena);
-                        let metric_node = match ctx.registry.process_metrics {
-                            Some(metrics) => {
-                                let token = metric_token.ok_or_else(|| {
-                                    anyhow::anyhow!("compiled inline process token is missing")
-                                })?;
-                                Some(metrics.select_node(token).ok_or_else(|| {
-                                    anyhow::anyhow!("compiled inline process token is out of range")
-                                })?)
-                            }
-                            None => None,
-                        };
-                        if let Some(node) = metric_node {
-                            node.counters.start();
-                        }
-                        let result = match metric_node {
-                            Some(node) => exec_process_body_with_metric_plan(
-                                body,
-                                &node.body_plan,
-                                current,
-                                ctx.registry,
-                                ctx.funcs,
-                                ctx.arena,
-                            ),
-                            None => {
-                                exec_process_body(body, current, ctx.registry, ctx.funcs, ctx.arena)
-                            }
-                        };
-                        match result {
-                            Ok(ExecResult::Continue(e)) => {
-                                if let Some(node) = metric_node {
-                                    node.counters.continued();
-                                }
-                                out.push_trace(|| TraceEntry {
-                                    stage: "process".into(),
-                                    label: "(inline)".into(),
-                                    detail: "ok".into(),
-                                });
-                                current = e;
-                            }
-                            Ok(ExecResult::Dropped) => {
-                                if let Some(node) = metric_node {
-                                    node.counters.dropped();
-                                }
-                                out.push_trace(|| TraceEntry {
-                                    stage: "process".into(),
-                                    label: "(inline)".into(),
-                                    detail: "dropped".into(),
-                                });
-                                return dropped();
-                            }
-                            Err(e) => {
-                                if let Some(node) = metric_node {
-                                    node.counters.errored();
-                                }
-                                tracing::warn!("inline process: {} — event routed to error_log", e);
-                                out.push_trace(|| TraceEntry {
-                                    stage: "process".into(),
-                                    label: "(inline)".into(),
-                                    detail: format!("error: {} (event → error_log)", e),
-                                });
-                                out.errored.push(ErroredEventContext::Process {
-                                    timestamp: chrono::Utc::now(),
-                                    pipeline: ctx.pipeline_name.to_string(),
-                                    site: "(inline)".to_string(),
-                                    reason: e.to_string(),
-                                    event: ProcessEvent::from_owned(&backup_view.to_owned()),
-                                });
-                                return Ok((None, PipelineTermination::Errored));
-                            }
-                        }
-                    }
-                }
-            }
-            cont(current)
-        }
-
-        PipelineStatement::Output(name) => {
-            trace!(target: "limpid::pipeline", "output → {}", name);
-            out.push_trace(|| TraceEntry {
-                stage: "output".into(),
-                label: format!("→ {}", name),
-                detail: String::new(),
-            });
-            // The queue transports a plain `OwnedEvent`; both memory
-            // and disk queues carry `Event` end-to-end and render
-            // runs consumer-side inside each sink's `Output::consume`.
-            // What the snapshot pushed here contains is decided by
-            // `OutputCapturePolicy`:
-            //
-            // - Memory queue + non-test-pipeline: the snapshot drops
-            //   the `workspace` (via `to_owned_without_workspace`).
-            //   The downstream sink reads `egress` (and, for `file`
-            //   and `kafka`, `source` / `received_at` / `source.ip`),
-            //   the DLQ path projects to `OutputEvent`'s five fields,
-            //   and the output-flavor `tap` strips `workspace` on the
-            //   emit side too — so nobody observes the missing
-            //   workspace.
-            // - Disk queue (or `--test-pipeline`): the snapshot keeps
-            //   the `workspace` (via `to_owned`). Disk queues need it
-            //   because the WAL persists the full `Event` JSON and
-            //   replay rehydrates it; `--test-pipeline` needs it
-            //   because its CLI display shows the snapshot verbatim.
-            //
-            // The live `event` is unchanged either way — any
-            // subsequent `if workspace.x == ...` gate at pipeline
-            // scope still sees the populated workspace on the
-            // borrowed view.
-            let snapshot = if ctx.output_capture.should_capture_workspace(name) {
-                event.to_owned()
-            } else {
-                event.to_owned_without_workspace()
-            };
-            let emitted_at = crate::time::UnixNanos::now();
-            if let Some(PipelineMetricStatement::Output(timer)) = metric_stmt {
-                timer.observe_between(ctx.dispatch_started_at, emitted_at);
-            } else if ctx.registry.process_metrics.is_some() {
-                bail!("compiled pipeline output metric entry is missing");
-            }
-            out.outputs
-                .push((name.clone(), QueuedEvent::new(snapshot, emitted_at)));
-            cont(event)
-        }
-
-        PipelineStatement::Drop => {
-            trace!(target: "limpid::pipeline", "drop");
-            out.push_trace(|| TraceEntry {
-                stage: "drop".into(),
-                label: String::new(),
-                detail: String::new(),
-            });
-            dropped()
-        }
-
-        PipelineStatement::Finish => {
-            trace!(target: "limpid::pipeline", "finish");
-            out.push_trace(|| TraceEntry {
-                stage: "finish".into(),
-                label: String::new(),
-                detail: String::new(),
-            });
-            finished()
-        }
-
-        PipelineStatement::If(if_chain) => {
-            match select_if_branch_with_ordinal(if_chain, |c| {
-                eval_expr(c, &event, ctx.funcs, ctx.arena)
-            })? {
-                Some(selection) => {
-                    let metric_body = match metric_stmt {
-                        Some(PipelineMetricStatement::If {
-                            branches,
-                            else_body,
-                        }) => match selection.ordinal {
-                            Some(ordinal) => branches.get(ordinal).map(Vec::as_slice),
-                            None => else_body.as_deref(),
-                        },
-                        _ => None,
-                    };
-                    exec_pipeline_branch_body(selection.body, metric_body, event, ctx, out)
-                }
-                None => cont(event),
-            }
-        }
-
-        PipelineStatement::Switch(discriminant, arms) => {
-            let disc_val = eval_expr(discriminant, &event, ctx.funcs, ctx.arena)?;
-            match select_switch_arm_with_ordinal(&disc_val, arms, |e| {
-                eval_expr(e, &event, ctx.funcs, ctx.arena)
-            })? {
-                Some((ordinal, body)) => {
-                    let metric_body = match metric_stmt {
-                        Some(PipelineMetricStatement::Switch(metrics)) => {
-                            metrics.get(ordinal).map(Vec::as_slice)
-                        }
-                        _ => None,
-                    };
-                    exec_pipeline_branch_body(body, metric_body, event, ctx, out)
-                }
-                None => cont(event),
-            }
-        }
-    }
-}
-
-fn exec_pipeline_branch_body<'bump>(
-    body: &[BranchBody],
-    metric_body: Option<&[PipelineMetricStatement]>,
-    mut event: BorrowedEvent<'bump>,
-    ctx: &PipelineExecCtx<'_, 'bump>,
-    out: &mut PipelineExecOut<'_>,
-) -> Result<(Option<BorrowedEvent<'bump>>, PipelineTermination)> {
-    if ctx.registry.process_metrics.is_some() {
-        let metric_body = metric_body
-            .ok_or_else(|| anyhow::anyhow!("compiled pipeline branch plan is missing"))?;
-        if metric_body.len() != body.len() {
-            bail!("compiled pipeline branch plan length does not match the selected body");
-        }
-    }
-    for (index, item) in body.iter().enumerate() {
-        let metric_stmt = match ctx.registry.process_metrics {
-            Some(_) => Some(
-                metric_body
-                    .and_then(|metrics| metrics.get(index))
-                    .ok_or_else(|| anyhow::anyhow!("compiled pipeline branch entry is missing"))?,
-            ),
-            None => None,
-        };
-        match item {
-            BranchBody::Pipeline(stmt) => {
-                match exec_pipeline_stmt(stmt, metric_stmt, event, ctx, out)? {
-                    (Some(e), _) => event = e,
-                    (None, term) => return Ok((None, term)),
-                }
-            }
-            BranchBody::Process(_) => {
-                bail!("process statement found in pipeline context")
-            }
-        }
-    }
-    Ok((Some(event), PipelineTermination::Finished))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
