+++ crates/limpid/src/runtime.rs
@@ -23,291 +23,14 @@ use crate::pipeline::CompiledConfig;
 use crate::queue::{self, QueueConfig, QueueSender};
 use crate::tap::TapRegistry;
 
-async fn shutdown_change_is_terminal(shutdown: &mut watch::Receiver<bool>) -> bool {
-    match shutdown.changed().await {
-        Ok(()) => *shutdown.borrow(),
-        Err(_) => true,
-    }
-}
+mod lifecycle;
+use lifecycle::*;
 
 pub struct Runtime {
     shutdown_tx: watch::Sender<bool>,
     handles: Vec<TrackedTask>,
     config_file: PathBuf,
     compiled_config: CompiledConfig,
-}
-
-const SHUTDOWN_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(10);
-const ABORT_JOIN_RESERVE: std::time::Duration = std::time::Duration::from_secs(1);
-
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-struct CleanupOutcome {
-    forced_abort: bool,
-    abort_safe_incomplete: usize,
-    must_join_exceeded_threshold: bool,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum TaskKind {
-    AbortSafe,
-    MustJoin,
-}
-
-fn output_task_kind(
-    output_type: &str,
-    queue_type: &queue::QueueType,
-    has_error_log: bool,
-    stdout_regular_file: bool,
-) -> TaskKind {
-    if output_type == "file"
-        || (output_type == "stdout" && stdout_regular_file)
-        || matches!(queue_type, queue::QueueType::Disk { .. })
-        || has_error_log
-    {
-        TaskKind::MustJoin
-    } else {
-        TaskKind::AbortSafe
-    }
-}
-
-fn pipeline_task_kind(has_error_log: bool, has_disk_output: bool) -> TaskKind {
-    if has_error_log || has_disk_output {
-        TaskKind::MustJoin
-    } else {
-        TaskKind::AbortSafe
-    }
-}
-
-fn input_task_kind(input_type: &str) -> TaskKind {
-    if input_type == "journal" {
-        TaskKind::MustJoin
-    } else {
-        TaskKind::AbortSafe
-    }
-}
-
-struct TrackedTask {
-    kind: TaskKind,
-    handle: Option<tokio::task::JoinHandle<()>>,
-}
-
-/// Owns every task created during startup until the runtime is fully committed.
-/// Any explicit startup error uses [`Self::rollback`]; cancellation or panic is
-/// covered by `Drop`, which transfers cleanup to the current Tokio runtime.
-struct StartupGuard {
-    shutdown_tx: Option<watch::Sender<bool>>,
-    handles: Vec<TrackedTask>,
-    cleanup_executor: tokio::runtime::Handle,
-    #[cfg(test)]
-    drop_cleanup_observer: Option<Arc<DropCleanupObserver>>,
-}
-
-impl StartupGuard {
-    fn new(shutdown_tx: watch::Sender<bool>) -> Self {
-        Self {
-            shutdown_tx: Some(shutdown_tx),
-            handles: Vec::new(),
-            cleanup_executor: tokio::runtime::Handle::current(),
-            #[cfg(test)]
-            drop_cleanup_observer: DROP_CLEANUP_RESULT.try_with(Arc::clone).ok(),
-        }
-    }
-
-    fn track(&mut self, kind: TaskKind, handle: tokio::task::JoinHandle<()>) {
-        self.handles.push(TrackedTask {
-            kind,
-            handle: Some(handle),
-        });
-    }
-
-    fn extend(
-        &mut self,
-        kind: TaskKind,
-        handles: impl IntoIterator<Item = tokio::task::JoinHandle<()>>,
-    ) {
-        self.handles
-            .extend(handles.into_iter().map(|handle| TrackedTask {
-                kind,
-                handle: Some(handle),
-            }));
-    }
-
-    async fn rollback(self, original: anyhow::Error) -> anyhow::Error {
-        self.rollback_with_timeout(original, SHUTDOWN_TIMEOUT).await
-    }
-
-    async fn rollback_with_timeout(
-        mut self,
-        original: anyhow::Error,
-        timeout: std::time::Duration,
-    ) -> anyhow::Error {
-        let shutdown_tx = self.shutdown_tx.take().expect("startup guard sender");
-        let mut handles = std::mem::take(&mut self.handles);
-        let cleanup = shutdown_tasks_with_timeout(&shutdown_tx, &mut handles, timeout).await;
-        drop(shutdown_tx);
-        if cleanup.abort_safe_incomplete != 0 {
-            original.context(format!(
-                "runtime startup rollback reached the hard cleanup deadline; {} abort-safe task(s) remain incomplete after abort",
-                cleanup.abort_safe_incomplete
-            ))
-        } else if cleanup.must_join_exceeded_threshold {
-            original.context(
-                "runtime startup rollback exceeded the 10s health threshold; must-join resource owners completed before return",
-            )
-        } else if cleanup.forced_abort {
-            original.context(
-                "runtime startup rollback exceeded the graceful phase; pending tasks were aborted and joined within the global cleanup deadline",
-            )
-        } else {
-            original.context("runtime startup rolled back all started tasks")
-        }
-    }
-
-    fn commit(mut self) -> (watch::Sender<bool>, Vec<TrackedTask>) {
-        let shutdown_tx = self.shutdown_tx.take().expect("startup guard sender");
-        let handles = std::mem::take(&mut self.handles);
-        (shutdown_tx, handles)
-    }
-}
-
-impl Drop for StartupGuard {
-    fn drop(&mut self) {
-        let Some(shutdown_tx) = self.shutdown_tx.take() else {
-            return;
-        };
-        let mut handles = std::mem::take(&mut self.handles);
-        let _ = shutdown_tx.send(true);
-        if handles.is_empty() {
-            #[cfg(test)]
-            if let Some(observer) = self.drop_cleanup_observer.take() {
-                observer.complete(CleanupOutcome::default());
-            }
-            return;
-        }
-        #[cfg(test)]
-        let observer = self.drop_cleanup_observer.take();
-        // The originating runtime must remain alive until this cleanup task
-        // completes. Dropping the runtime itself cancels all tasks by Tokio's
-        // contract; no handle can extend an executor beyond that boundary.
-        self.cleanup_executor.spawn(async move {
-            let outcome = shutdown_tasks(&shutdown_tx, &mut handles).await;
-            #[cfg(test)]
-            if let Some(observer) = observer {
-                observer.complete(outcome);
-            }
-            #[cfg(not(test))]
-            let _ = outcome;
-        });
-    }
-}
-
-fn record_join_result(result: std::result::Result<(), tokio::task::JoinError>) {
-    if let Err(error) = result
-        && error.is_panic()
-    {
-        error!("task panicked during shutdown: {error}");
-    }
-}
-
-async fn shutdown_tasks(
-    shutdown_tx: &watch::Sender<bool>,
-    handles: &mut [TrackedTask],
-) -> CleanupOutcome {
-    shutdown_tasks_with_timeout(shutdown_tx, handles, SHUTDOWN_TIMEOUT).await
-}
-
-async fn shutdown_tasks_with_timeout(
-    shutdown_tx: &watch::Sender<bool>,
-    handles: &mut [TrackedTask],
-    total_timeout: std::time::Duration,
-) -> CleanupOutcome {
-    let _ = shutdown_tx.send(true);
-    let started = tokio::time::Instant::now();
-    let overall_deadline = started + total_timeout;
-    let abort_reserve = ABORT_JOIN_RESERVE.min(total_timeout / 2);
-    let graceful_deadline = overall_deadline - abort_reserve;
-
-    for task in handles.iter_mut() {
-        let Some(handle) = task.handle.as_mut() else {
-            continue;
-        };
-        match tokio::time::timeout_at(graceful_deadline, handle).await {
-            Ok(result) => {
-                record_join_result(result);
-                task.handle = None;
-            }
-            Err(_) => break,
-        }
-    }
-
-    let forced_abort = handles
-        .iter()
-        .any(|task| task.kind == TaskKind::AbortSafe && task.handle.is_some());
-    for task in handles.iter() {
-        if task.kind == TaskKind::AbortSafe
-            && let Some(handle) = &task.handle
-        {
-            handle.abort();
-        }
-    }
-
-    for task in handles.iter_mut() {
-        if task.kind != TaskKind::AbortSafe {
-            continue;
-        }
-        let Some(handle) = task.handle.as_mut() else {
-            continue;
-        };
-        match tokio::time::timeout_at(overall_deadline, handle).await {
-            Ok(result) => {
-                record_join_result(result);
-                task.handle = None;
-            }
-            Err(_) => break,
-        }
-    }
-
-    // Consume MustJoin handles that completed by the health threshold before
-    // deciding whether the threshold was exceeded. This also preserves the
-    // one-poll ownership invariant for every completed handle.
-    for task in handles.iter_mut() {
-        if task.kind == TaskKind::MustJoin
-            && task
-                .handle
-                .as_ref()
-                .is_some_and(|handle| handle.is_finished())
-            && let Some(handle) = task.handle.take()
-        {
-            record_join_result(handle.await);
-        }
-    }
-
-    let abort_safe_incomplete = handles
-        .iter()
-        .filter(|task| task.kind == TaskKind::AbortSafe && task.handle.is_some())
-        .count();
-    let must_join_exceeded_threshold = handles
-        .iter()
-        .any(|task| task.kind == TaskKind::MustJoin && task.handle.is_some());
-
-    // Must-join tasks own durability or disposition state. The 10-second
-    // deadline is a health threshold for them, never permission to abort or
-    // detach. Await every remaining owner before returning.
-    for task in handles.iter_mut() {
-        if task.kind != TaskKind::MustJoin {
-            continue;
-        }
-        if let Some(handle) = task.handle.as_mut() {
-            record_join_result(handle.await);
-            task.handle = None;
-        }
-    }
-
-    CleanupOutcome {
-        forced_abort,
-        abort_safe_incomplete,
-        must_join_exceeded_threshold,
-    }
 }
 
 #[cfg(test)]
@@ -362,1248 +85,11 @@ struct PostOutputFailure {
     enqueue_probe: bool,
 }
 
-async fn post_output_preactivation_failure(_sender: &mut QueueSender) -> Result<()> {
-    #[cfg(test)]
-    if let Some(failure) = POST_OUTPUT_PREACTIVATION_FAILURE
-        .try_with(|slot| slot.borrow_mut().take())
-        .ok()
-        .flatten()
-    {
-        if failure.enqueue_probe {
-            _sender
-                .send(crate::event::QueuedEvent::new(
-                    Event::new(
-                        bytes::Bytes::from_static(b"startup-rollback-probe"),
-                        "127.0.0.1:1".parse().expect("static probe address"),
-                    ),
-                    crate::time::UnixNanos::now(),
-                ))
-                .await
-                .context("failed to enqueue startup rollback probe")?;
-        }
-        failure.reached.notify_one();
-        match failure.mode {
-            PostOutputFailureMode::Error => {
-                anyhow::bail!("test failpoint: post-output preactivation failure");
-            }
-            PostOutputFailureMode::Park => std::future::pending::<()>().await,
-        }
-    }
-    Ok(())
-}
-
-fn late_post_listener_failure() -> Result<()> {
-    #[cfg(test)]
-    if LATE_POST_LISTENER_FAILURE
-        .try_with(|armed| armed.replace(false))
-        .unwrap_or(false)
-    {
-        anyhow::bail!("test failpoint: late post-listener startup failure");
-    }
-    Ok(())
-}
-
-fn configured_ltp_peer_ids(config: &CompiledConfig) -> Result<BTreeSet<String>> {
-    let mut peers = BTreeSet::new();
-    for (kind, name, properties) in config
-        .inputs
-        .iter()
-        .map(|(name, def)| ("input", name, &def.properties))
-        .chain(
-            config
-                .outputs
-                .iter()
-                .map(|(name, def)| ("output", name, &def.properties)),
-        )
-        .filter(|(_, _, properties)| properties.type_name() == "ltp")
-    {
-        for peer in properties
-            .user_properties()
-            .iter()
-            .filter_map(|property| match property {
-                Property::Block {
-                    key, properties, ..
-                } if key == "peer" => Some(properties.as_slice()),
-                _ => None,
-            })
-        {
-            let node_id = props::get_string(peer, "node_id").ok_or_else(|| {
-                anyhow::anyhow!("{kind} '{name}': peer node_id requires a string value")
-            })?;
-            peers.insert(node_id);
-        }
-    }
-    Ok(peers)
-}
-
-impl Runtime {
-    pub async fn start(config: CompiledConfig, config_file: PathBuf) -> Result<Self> {
-        Self::start_with_registry(config, config_file, Arc::new(Registry::new())).await
-    }
-
-    pub(crate) async fn start_with_registry(
-        config: CompiledConfig,
-        config_file: PathBuf,
-        metrics_registry: Arc<Registry>,
-    ) -> Result<Self> {
-        Self::start_with_registry_and_node_id_resolver(
-            config,
-            config_file,
-            metrics_registry,
-            || Ok(gethostname::gethostname().to_string_lossy().into_owned()),
-        )
-        .await
-    }
-
-    pub(crate) async fn start_with_registry_and_node_id_resolver<F>(
-        config: CompiledConfig,
-        config_file: PathBuf,
-        metrics_registry: Arc<Registry>,
-        resolve_hostname: F,
-    ) -> Result<Self>
-    where
-        F: FnOnce() -> Result<String>,
-    {
-        let (shutdown_tx, shutdown_rx) = watch::channel(false);
-
-        let mut registry = ModuleRegistry::new();
-        modules::register_builtins(&mut registry);
-        // Future: dynamic plugin loading from /etc/limpid/plugins/
-
-        init_geoip(&config);
-        let table_store = init_tables(&config)?;
-
-        let mut func_registry = FunctionRegistry::new();
-        crate::functions::register_builtins(&mut func_registry, table_store);
-        crate::functions::register_user_functions(&mut func_registry, &config);
-        let func_registry = Arc::new(func_registry);
-
-        config.validate()?;
-        let ltp_peer_ids = configured_ltp_peer_ids(&config)?;
-        let ltp_metrics = if ltp_peer_ids.is_empty() {
-            None
-        } else {
-            Some(LtpMetrics::register(&metrics_registry, &ltp_peer_ids)?)
-        };
-        let ltp_node_key = config
-            .node_key
-            .as_deref()
-            .map(Path::new)
-            .map(crate::ltp::load_node_key)
-            .transpose()?
-            .map(Arc::new);
-        let node_id = match &config.node_id {
-            Some(node_id) => node_id.clone(),
-            None => resolve_hostname()?,
-        };
-        crate::metrics::register_build_info(
-            &metrics_registry,
-            env!("CARGO_PKG_VERSION"),
-            &node_id,
-        )?;
-        let registry = Arc::new(registry);
-
-        let tap = TapRegistry::new();
-
-        // Optional dead-letter queue for events that fail in `process`
-        // or that an output drops after exhausting retries
-        // (retry-exhausted recovery). `control { error_log "..." }`
-        // opts in to file-based recovery; when unset, every emission
-        // site delegates to `emit_dlq_tracing_fallback`, which
-        // enforces the operator's `error_log_fallback` ladder —
-        // payload-free summary by default (`Off`), structured
-        // metadata on `Meta`, or full JSONL via `event_record`
-        // on `Full`. Pipeline-side and sink-side paths share the
-        // same helper, so the ladder shape is identical across
-        // both surfaces. The path is validated at startup (parent
-        // dir reachable) so operator typos surface before the
-        // first failure event.
-        //
-        // Built *before* outputs are constructed so each batched output
-        // (`http`, `otlp_http`, `otlp_grpc`) receives the handle via
-        // its constructor — no post-construction setter, no interior
-        // mutability. Non-batched outputs ignore the parameter; the
-        // queue consumer hands `error_log` to `run_queue_consumer`,
-        // which routes the retry-exhausted payload to the DLQ once
-        // each handle resolves.
-        let error_log_path = config
-            .global_blocks
-            .get("control")
-            .and_then(|p| props::get_string(p, "error_log"));
-        let error_log = match error_log_path {
-            Some(p) => {
-                let writer = crate::error_log::ErrorLogWriter::new(PathBuf::from(p));
-                writer.validate_at_startup().await?;
-                Some(Arc::new(writer))
-            }
-            None => None,
-        };
-
-        // Fallback policy for the tracing line when `error_log` write
-        // fails or is unset. Parsed here so invalid values fail
-        // daemon startup (matching `--check`'s config-time refusal).
-        let error_log_fallback = match config
-            .global_blocks
-            .get("control")
-            .and_then(|p| props::get_string(p, "error_log_fallback"))
-        {
-            Some(s) => crate::error_log::ErrorLogFallback::parse(&s)
-                .map_err(|e| anyhow::anyhow!("{}", e))?,
-            None => crate::error_log::ErrorLogFallback::default(),
-        };
-        if error_log.is_none()
-            && error_log_fallback != crate::error_log::ErrorLogFallback::default()
-        {
-            warn!(
-                "control.error_log_fallback = \"{}\" is set but control.error_log is unset — \
-                 tracing fallback stays payload-free because no durable DLQ was requested; \
-                 either set control.error_log to opt into the fallback, or remove \
-                 control.error_log_fallback to silence this warning",
-                error_log_fallback.as_str(),
-            );
-        }
-
-        // Single bundle threaded into every Input/Output factory. Future
-        // build-time dependencies (transport-key registry, metrics hooks)
-        // land as new fields on this struct rather than as new parameters.
-        let build_ctx = crate::modules::BuildContext {
-            funcs: Arc::clone(&func_registry),
-            metrics: Arc::clone(&metrics_registry),
-            error_log: error_log.as_ref().map(Arc::clone),
-            error_log_fallback,
-            shutdown_signal: shutdown_rx.clone(),
-            ltp_node_id: Some(Arc::<str>::from(node_id.clone())),
-            ltp_node_key,
-            ltp_metrics,
-        };
-
-        // Complete every deterministic plan named by the startup contract
-        // before an output factory can acquire a resource.
-        let compiled_config = config.clone();
-        let control_path = config
-            .global_blocks
-            .get("control")
-            .and_then(|p| props::get_string(p, "socket"));
-        crate::control::validate_control_socket_parent(control_path.as_deref())?;
-        let stdout_regular_file = if config
-            .outputs
-            .values()
-            .any(|output| output.properties.type_name() == "stdout")
-        {
-            crate::modules::output::stdout::stdout_is_regular_file()
-                .context("failed to classify stdout backend")?
-        } else {
-            false
-        };
-
-        // Group pipelines by input and compile every process-metric plan.
-        let mut input_pipelines: HashMap<String, Vec<Arc<PipelineWorker>>> = HashMap::new();
-        for pipeline_def in config.pipelines.values() {
-            let worker = Arc::new(PipelineWorker::new_with_process_metrics(
-                pipeline_def.clone(),
-                &config.processes,
-                &metrics_registry,
-            )?);
-            for input_name in get_pipeline_inputs(pipeline_def) {
-                input_pipelines
-                    .entry(input_name.clone())
-                    .or_default()
-                    .push(Arc::clone(&worker));
-            }
-        }
-
-        let mut input_queue_sizes = HashMap::new();
-        let mut prepared_ltp_inputs = HashMap::new();
-        for input_name in input_pipelines.keys() {
-            let input_def = config
-                .inputs
-                .get(input_name)
-                .ok_or_else(|| anyhow::anyhow!("input '{}' not found", input_name))?;
-            let queue_size =
-                props::get_positive_int(input_def.properties.user_properties(), "queue_size")?
-                    .unwrap_or(4096) as usize;
-            input_queue_sizes.insert(input_name.clone(), queue_size);
-            if input_def.properties.type_name() == "ltp" {
-                let input = modules::input::ltp::LtpInput::build(
-                    input_name,
-                    &input_def.properties,
-                    &build_ctx,
-                )
-                .with_context(|| format!("failed to create input '{input_name}'"))?;
-                prepared_ltp_inputs.insert(input_name.clone(), input);
-            }
-        }
-
-        // Tap registration does not create runtime actors or output resources.
-        for input_name in input_pipelines.keys() {
-            tap.register(&format!("input {}", input_name)).await;
-        }
-        for proc_name in config.processes.keys() {
-            tap.register(&format!("process {}", proc_name)).await;
-        }
-
-        // From this point on, every acquired output resource is transactionally
-        // owned. Each real queue consumer is spawned and registered immediately
-        // after its output factory succeeds, before the next await or fallible
-        // edge.
-        let mut startup_guard = StartupGuard::new(shutdown_tx);
-
-        // --- 1. Create outputs (each output owns its own OutputMetrics) ---
-        let mut output_senders: HashMap<String, QueueSender> = HashMap::new();
-        // Populated in the loop below alongside the queue creation so
-        // the same `QueueConfig` decides which set of outputs need a
-        // workspace-carrying snapshot at `output` statement time. See
-        // `PipelineContext::disk_outputs` for the runtime contract.
-        let mut disk_outputs: HashSet<String> = HashSet::new();
-
-        for (name, output_def) in &config.outputs {
-            let queue_config = match QueueConfig::from_output_properties(
-                name,
-                output_def.properties.user_properties(),
-            ) {
-                Ok(config) => config,
-                Err(error) => return Err(startup_guard.rollback(error).await),
-            };
-            let output_task_kind = output_task_kind(
-                output_def.properties.type_name(),
-                &queue_config.queue_type,
-                error_log.is_some(),
-                stdout_regular_file,
-            );
-            if matches!(queue_config.queue_type, queue::QueueType::Disk { .. }) {
-                disk_outputs.insert(name.clone());
-            }
-            // Retry config is parsed by each output's `from_properties`
-            // (outputs own retry + DLQ). The runtime no longer needs a
-            // copy here.
-            let (mut sender, receiver) = match queue::create_queue(name.clone(), queue_config) {
-                Ok(queue) => queue,
-                Err(error) => return Err(startup_guard.rollback(error).await),
-            };
-
-            // `output_def.properties` is a `ModuleProperties`: it carries the
-            // resolved `type` already, so `create_output` doesn't take a
-            // separate type_name argument (and can't be passed one — the
-            // strip is the whole point). `BuildContext` carries `funcs` and
-            // the optional `error_log` so outputs can stash them at
-            // construction time.
-            let created = match registry
-                .create_output(name, &output_def.properties, &build_ctx)
-                .with_context(|| format!("failed to create output '{name}'"))
-            {
-                Ok(created) => created,
-                Err(error) => return Err(startup_guard.rollback(error).await),
-            };
-
-            // Attach metrics so QueueSender::send counts events_received.
-            sender.attach_metrics(Arc::clone(&created.metrics));
-            let output_metrics = Arc::clone(&created.metrics);
-            let shutdown = shutdown_rx.clone();
-            let tap_clone = tap.clone();
-            let error_log_for_consumer = error_log.as_ref().map(Arc::clone);
-            #[cfg(test)]
-            let completion_observer = STARTUP_TASK_COMPLETIONS.try_with(Arc::clone).ok();
-            startup_guard.track(
-                output_task_kind,
-                tokio::spawn(async move {
-                    queue::run_queue_consumer(
-                        receiver,
-                        created.output,
-                        Some(tap_clone),
-                        output_metrics,
-                        error_log_for_consumer,
-                        shutdown,
-                    )
-                    .await;
-                    #[cfg(test)]
-                    if let Some(observer) = completion_observer {
-                        observer
-                            .output_queues
-                            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                    }
-                }),
-            );
-            if let Err(error) = post_output_preactivation_failure(&mut sender).await {
-                return Err(startup_guard.rollback(error).await);
-            }
-            output_senders.insert(name.clone(), sender);
-            tap.register(&format!("output {}", name)).await;
-        }
-
-        let output_senders = Arc::new(output_senders);
-        let has_disk_output = !disk_outputs.is_empty();
-        let disk_outputs = Arc::new(disk_outputs);
-
-        let config = Arc::new(config);
-
-        // --- 3. Start inputs (each input owns its own InputMetrics) ---
-
-        let mut input_senders: HashMap<
-            String,
-            (mpsc::Sender<Event>, Arc<crate::metrics::InputMetrics>),
-        > = HashMap::new();
-        let mut ltp_inputs = Vec::new();
-
-        for (input_name, pipelines) in input_pipelines {
-            let input_def = config.inputs.get(&input_name).expect("input preflight");
-            let queue_size = input_queue_sizes
-                .remove(&input_name)
-                .expect("input queue-size preflight");
-            let (event_tx, event_rx) = mpsc::channel::<Event>(queue_size);
-            let input_queue_timer =
-                match crate::metrics::InputQueueTimer::register(&metrics_registry, &input_name) {
-                    Ok(timer) => timer,
-                    Err(error) => return Err(startup_guard.rollback(error.into()).await),
-                };
-
-            // Pipeline workers subscribed to this input. A pipeline with fan-in
-            // (`input a, b;`) appears in the worker list of both inputs — its
-            // merge semantics is implicit: two dispatcher tasks feeding the
-            // same `PipelineWorker`, serialized through its own `run_pipeline`
-            // call per event. No ordering guarantee between inputs.
-            let workers: Arc<Vec<Arc<PipelineWorker>>> = Arc::new(pipelines);
-            let ctx = PipelineContext {
-                output_senders: Arc::clone(&output_senders),
-                disk_outputs: Arc::clone(&disk_outputs),
-                config: Arc::clone(&config),
-                funcs: Arc::clone(&func_registry),
-                tap: tap.clone(),
-                error_log: error_log.as_ref().map(Arc::clone),
-                error_log_fallback,
-            };
-            let iname = input_name.clone();
-            let shutdown_for_worker = shutdown_rx.clone();
-            let sender_for_inject = event_tx.clone();
-            #[cfg(test)]
-            let completion_observer = STARTUP_TASK_COMPLETIONS.try_with(Arc::clone).ok();
-            #[cfg(test)]
-            let wal_barrier = queue::current_test_wal_barrier();
-            let pipeline_task_kind = pipeline_task_kind(error_log.is_some(), has_disk_output);
-            startup_guard.track(
-                pipeline_task_kind,
-                tokio::spawn(async move {
-                    #[cfg(test)]
-                    if let Some(barrier) = wal_barrier {
-                        queue::with_test_wal_barrier(
-                            barrier,
-                            run_pipeline_workers(
-                                event_rx,
-                                &workers,
-                                &ctx,
-                                &iname,
-                                &input_queue_timer,
-                                shutdown_for_worker,
-                            ),
-                        )
-                        .await;
-                    } else {
-                        run_pipeline_workers(
-                            event_rx,
-                            &workers,
-                            &ctx,
-                            &iname,
-                            &input_queue_timer,
-                            shutdown_for_worker,
-                        )
-                        .await;
-                    }
-                    #[cfg(not(test))]
-                    run_pipeline_workers(
-                        event_rx,
-                        &workers,
-                        &ctx,
-                        &iname,
-                        &input_queue_timer,
-                        shutdown_for_worker,
-                    )
-                    .await;
-                    #[cfg(test)]
-                    if let Some(observer) = completion_observer {
-                        observer
-                            .pipelines
-                            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                    }
-                }),
-            );
-
-            if input_def.properties.type_name() == "ltp" {
-                let input = prepared_ltp_inputs
-                    .remove(&input_name)
-                    .expect("LTP input preflight");
-                input_senders.insert(input_name.clone(), (sender_for_inject, input.metrics()));
-                ltp_inputs.push((input, event_tx));
-                continue;
-            }
-
-            // Input — registry builds, spawns, and returns metrics handle.
-            // `input_def.properties` carries the resolved `type`; no separate
-            // type_name argument needed (see ModuleProperties rationale).
-            let created = match registry.create_input(
-                &input_name,
-                &input_def.properties,
-                &build_ctx,
-                event_tx,
-                shutdown_rx.clone(),
-            ) {
-                Ok(c) => c,
-                Err(e) => {
-                    let error = e.context(format!("failed to start input '{input_name}'"));
-                    return Err(startup_guard.rollback(error).await);
-                }
-            };
-            input_senders.insert(
-                input_name.clone(),
-                (sender_for_inject, Arc::clone(&created.metrics)),
-            );
-            let input_task_kind = input_task_kind(input_def.properties.type_name());
-            startup_guard.track(input_task_kind, created.handle);
-            match created.startup.await {
-                Ok(Ok(())) => {}
-                Ok(Err(error)) => {
-                    let error = error.context(format!("failed to start input '{input_name}'"));
-                    return Err(startup_guard.rollback(error).await);
-                }
-                Err(_) => {
-                    let error = anyhow::anyhow!(
-                        "input '{input_name}' task exited before startup readiness"
-                    );
-                    return Err(startup_guard.rollback(error).await);
-                }
-            }
-        }
-
-        let ltp_handles =
-            match modules::input::ltp::start_listener_groups(ltp_inputs, shutdown_rx.clone()).await
-            {
-                Ok(handles) => handles,
-                Err(error) => {
-                    return Err(startup_guard
-                        .rollback(error.context("failed to start LTP listener groups"))
-                        .await);
-                }
-            };
-        startup_guard.extend(TaskKind::AbortSafe, ltp_handles);
-
-        if let Err(error) = late_post_listener_failure() {
-            return Err(startup_guard.rollback(error).await);
-        }
-
-        // --- 4. Start control socket (after all metrics are registered) ---
-        let started_at = std::time::Instant::now();
-        let control = ControlServer::new(
-            control_path,
-            tap.clone(),
-            Arc::clone(&metrics_registry),
-            Arc::clone(&config),
-            input_senders,
-            Arc::clone(&output_senders),
-            started_at,
-        );
-        let s = shutdown_rx.clone();
-        let (control_startup_tx, control_startup_rx) = tokio::sync::oneshot::channel();
-        startup_guard.track(
-            TaskKind::AbortSafe,
-            tokio::spawn(async move {
-                control.run(s, Some(control_startup_tx)).await;
-            }),
-        );
-        match control_startup_rx.await {
-            Ok(Ok(())) => {}
-            Ok(Err(diagnostic)) => {
-                return Err(startup_guard.rollback(anyhow::anyhow!(diagnostic)).await);
-            }
-            Err(_) => {
-                return Err(startup_guard
-                    .rollback(anyhow::anyhow!(
-                        "control task exited before startup readiness"
-                    ))
-                    .await);
-            }
-        }
-
-        let (shutdown_tx, handles) = startup_guard.commit();
-        info!("limpid daemon started");
-        Ok(Self {
-            shutdown_tx,
-            handles,
-            config_file,
-            compiled_config,
-        })
-    }
-
-    pub fn config_file(&self) -> &Path {
-        &self.config_file
-    }
-
-    pub fn compiled_config(&self) -> CompiledConfig {
-        self.compiled_config.clone()
-    }
-
-    pub async fn shutdown(self) {
-        info!(
-            "initiating graceful shutdown (timeout: {}s)",
-            SHUTDOWN_TIMEOUT.as_secs()
-        );
-        let mut handles = self.handles;
-        let cleanup = shutdown_tasks(&self.shutdown_tx, &mut handles).await;
-        if cleanup.abort_safe_incomplete != 0 {
-            error!(
-                "shutdown reached the {}s hard deadline — {} abort-safe task(s) remain incomplete after abort",
-                SHUTDOWN_TIMEOUT.as_secs(),
-                cleanup.abort_safe_incomplete,
-            );
-        } else if cleanup.must_join_exceeded_threshold {
-            warn!(
-                "shutdown exceeded the {}s health threshold; must-join resource owners completed before return",
-                SHUTDOWN_TIMEOUT.as_secs(),
-            );
-        } else if cleanup.forced_abort {
-            warn!("shutdown exceeded the graceful phase; pending tasks were aborted and joined");
-        } else {
-            info!("shutdown complete");
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Global subsystem initialization
-// ---------------------------------------------------------------------------
-
-fn init_geoip(config: &CompiledConfig) {
-    let db_path = config
-        .global_blocks
-        .get("geoip")
-        .and_then(|p| props::get_string(p, "database"))
-        .map(PathBuf::from);
-    crate::functions::geoip::init(db_path.as_ref());
-}
-
-pub(crate) fn init_tables(config: &CompiledConfig) -> Result<crate::functions::table::TableStore> {
-    use crate::dsl::ast::Property;
-    use crate::functions::table::{TableConfig, TableStore};
-    use std::time::Duration;
-
-    let mut configs = Vec::new();
-
-    if let Some(props) = config.global_blocks.get("table") {
-        for prop in props {
-            if let Property::Block {
-                key: table_name,
-                properties: inner_props,
-                ..
-            } = prop
-            {
-                let load_path = props::get_string(inner_props, "load").map(PathBuf::from);
-                let max = props::get_positive_int(inner_props, "max")?.map(|n| n as usize);
-                let ttl = props::get_positive_int(inner_props, "ttl")?.map(Duration::from_secs);
-
-                configs.push(TableConfig {
-                    name: table_name.clone(),
-                    max,
-                    default_ttl: ttl,
-                    load_path,
-                });
-            }
-        }
-    }
-
-    TableStore::from_configs(configs)
-}
-
-// ---------------------------------------------------------------------------
-// Pipeline context — shared references for pipeline execution
-// ---------------------------------------------------------------------------
-
-struct PipelineContext {
-    output_senders: Arc<HashMap<String, QueueSender>>,
-    /// Names of outputs whose queue backend is disk-based. Precomputed
-    /// at startup from each output's parsed `QueueConfig` and handed
-    /// to `run_pipeline` per event via
-    /// [`crate::pipeline::OutputCapturePolicy::DiskOnly`], which uses
-    /// it to decide per-output whether to keep the workspace on the
-    /// snapshot pushed to the queue. Disk-backed queues need the
-    /// workspace because the WAL persists the full `Event` JSON and
-    /// replay rehydrates it; memory queues drop it because no
-    /// downstream reader (sink, DLQ record, output tap) touches
-    /// workspace on that path.
-    disk_outputs: Arc<HashSet<String>>,
-    config: Arc<CompiledConfig>,
-    funcs: Arc<FunctionRegistry>,
-    tap: TapRegistry,
-    /// Dead-letter queue writer used for: (1) `process` runtime
-    /// errors, (2) output retry-exhausted payloads, (3)
-    /// batched-output shutdown-flush leftovers, (4) runtime-side
-    /// enqueue failures (queue closed, disk write error, unknown
-    /// output). `None` when `control { error_log }` is unset — on
-    /// that path every emission site delegates to
-    /// [`crate::modules::emit_dlq_tracing_fallback`], which enforces
-    /// the operator's `error_log_fallback` ladder policy: payload-
-    /// free summary by default (`Off`), structured metadata on
-    /// opt-in (`Meta`), or the pre-ladder full-JSONL shape on
-    /// explicit `Full` opt-in. See the ladder documentation in
-    /// `docs/src/operations/error-log.md` for the confidentiality
-    /// rationale.
-    error_log: Option<Arc<crate::error_log::ErrorLogWriter>>,
-    /// Operator-selected confidentiality policy for the tracing-side
-    /// fallback line. Threaded through `write_errored_to_dlq` so the
-    /// pipeline-side runtime error surface obeys the same ladder as
-    /// the sink-side DLQ paths.
-    error_log_fallback: crate::error_log::ErrorLogFallback,
-}
-
-// ---------------------------------------------------------------------------
-// Pipeline worker — owns its own metrics via HasMetrics
-// ---------------------------------------------------------------------------
-
-struct PipelineWorker {
-    def: PipelineDef,
-    metrics: Arc<PipelineMetrics>,
-    process_metrics: Option<Arc<crate::pipeline::PipelineProcessMetrics>>,
-    #[cfg(test)]
-    serial_test_gate: Option<Arc<tokio::sync::Barrier>>,
-}
-
-impl PipelineWorker {
-    #[cfg(test)]
-    fn new(def: PipelineDef, registry: &Registry) -> Result<Self, crate::metrics::MetricsError> {
-        let metrics = PipelineMetrics::register(registry, &def.name)?;
-        Ok(Self {
-            def,
-            metrics,
-            process_metrics: None,
-            serial_test_gate: None,
-        })
-    }
-
-    fn new_with_process_metrics(
-        def: PipelineDef,
-        processes: &HashMap<String, crate::dsl::ast::ProcessDef>,
-        registry: &Registry,
-    ) -> anyhow::Result<Self> {
-        let metrics = PipelineMetrics::register(registry, &def.name)?;
-        let process_metrics = Arc::new(crate::pipeline::PipelineProcessMetrics::register(
-            &def, processes, registry,
-        )?);
-        Ok(Self {
-            def,
-            metrics,
-            process_metrics: Some(process_metrics),
-            #[cfg(test)]
-            serial_test_gate: None,
-        })
-    }
-
-    #[cfg(test)]
-    fn compile_process_metric_plan_for_testing(
-        def: &PipelineDef,
-        processes: &HashMap<String, crate::dsl::ast::ProcessDef>,
-        registry: &Registry,
-    ) -> Result<CompiledProcessMetricPlan, crate::metrics::MetricsError> {
-        Ok(CompiledProcessMetricPlan {
-            pipeline_metrics: PipelineMetrics::register(registry, &def.name)?,
-            process_metrics: crate::pipeline::PipelineProcessMetrics::compile_raw(
-                def, processes, registry,
-            )?,
-        })
-    }
-
-    #[cfg(test)]
-    fn new_with_compiled_process_metrics_for_testing(
-        def: PipelineDef,
-        processes: &HashMap<String, crate::dsl::ast::ProcessDef>,
-        plan: CompiledProcessMetricPlan,
-    ) -> anyhow::Result<Self> {
-        Ok(Self {
-            process_metrics: Some(Arc::new(plan.process_metrics.validate(&def, processes)?)),
-            def,
-            metrics: plan.pipeline_metrics,
-            serial_test_gate: None,
-        })
-    }
-}
-
-#[cfg(test)]
-struct CompiledProcessMetricPlan {
-    pipeline_metrics: Arc<PipelineMetrics>,
-    process_metrics: crate::pipeline::RawPipelineProcessMetrics,
-}
-
-#[cfg(test)]
-impl CompiledProcessMetricPlan {
-    fn process_metrics_mut_for_testing(
-        &mut self,
-    ) -> &mut crate::pipeline::RawPipelineProcessMetrics {
-        &mut self.process_metrics
-    }
-
-    fn root_token_for_testing(&self, step: usize) -> Option<usize> {
-        self.process_metrics.root_token_for_testing(step)
-    }
-
-    fn child_token_for_testing(&self, parent: usize, ordinal: usize) -> Option<usize> {
-        self.process_metrics
-            .child_token_for_testing(parent, ordinal)
-    }
-
-    fn metric_node_selection_trap_for_testing(&self) -> crate::pipeline::MetricNodeSelectionTrap {
-        self.process_metrics
-            .metric_node_selection_trap_for_testing()
-    }
-}
-
-impl HasMetrics for PipelineWorker {
-    type Stats = PipelineMetrics;
-    fn metrics(&self) -> Arc<PipelineMetrics> {
-        Arc::clone(&self.metrics)
-    }
-}
-
-async fn run_pipeline_workers(
-    mut event_rx: mpsc::Receiver<Event>,
-    workers: &[Arc<PipelineWorker>],
-    ctx: &PipelineContext,
-    input_name: &str,
-    input_queue_timer: &crate::metrics::InputQueueTimer,
-    mut shutdown: tokio::sync::watch::Receiver<bool>,
-) {
-    info!(
-        "pipeline worker for input '{}' started ({} pipeline(s))",
-        input_name,
-        workers.len()
-    );
-
-    let input_tap_key = format!("input {}", input_name);
-
-    // Per-worker arena. Bump-allocator state is owned by this task and
-    // recycled via `Bump::reset()` after each event, so the underlying
-    // chunk-group is malloc'd once at task startup and stays around as
-    // long as one event's working set fits in the initial chunk.
-    // 1 MiB comfortably fits the heaviest realistic OCSF compose (the
-    // D pipeline's measured workspace tree under load is well under
-    // 256 KiB). Without enough headroom, bumpalo grows the chunk
-    // chain mid-event and `reset` then frees the excess — defeating
-    // the whole purpose of pooling.
-    let mut bump = bumpalo::Bump::with_capacity(1024 * 1024);
-
-    loop {
-        let event = tokio::select! {
-            biased;
-
-            terminal = shutdown_change_is_terminal(&mut shutdown) => {
-                if terminal {
-                    // Close the receiver first, then drain with
-                    // `recv().await` until `None`. The previous
-                    // `try_recv()` snapshot loop raced with input
-                    // tasks that had reserved an mpsc permit but not
-                    // yet written the value — the loop could exit
-                    // observing an empty channel while a permit-holder
-                    // was mid-write, silently dropping that event.
-                    // `close()` + `recv-until-None` is the tokio-
-                    // documented contract for reading every value that
-                    // was already sent or is being sent by an
-                    // outstanding permit-holder before terminating.
-                    // Any input-side `send` that races the close and
-                    // has not yet reserved a permit wakes with
-                    // `SendError`, surfaced by the input tasks
-                    // themselves (see `run_input`).
-                    event_rx.close();
-                    while let Some(event) = event_rx.recv().await {
-                        process_event(
-                            &event,
-                            workers,
-                            ctx,
-                            &input_tap_key,
-                            input_queue_timer,
-                            &mut bump,
-                        )
-                        .await;
-                        bump.reset();
-                    }
-                    break;
-                }
-                continue;
-            }
-
-            event = event_rx.recv() => {
-                match event {
-                    Some(e) => e,
-                    None => break,
-                }
-            }
-        };
-        process_event(
-            &event,
-            workers,
-            ctx,
-            &input_tap_key,
-            input_queue_timer,
-            &mut bump,
-        )
-        .await;
-        bump.reset();
-    }
-
-    info!("pipeline worker for input '{}' stopped", input_name);
-}
-
-#[cfg(test)]
-async fn run_pipeline_with_outputs(
-    pipeline: &PipelineDef,
-    event: &Event,
-    ctx: &PipelineContext,
-    bump: &mut bumpalo::Bump,
-) -> Result<crate::pipeline::PipelineRunResult> {
-    run_pipeline_with_outputs_inner(
-        pipeline,
-        None,
-        event,
-        ctx,
-        bump,
-        crate::time::UnixNanos::now(),
-    )
-    .await
-}
-
-async fn run_pipeline_with_outputs_inner(
-    pipeline: &PipelineDef,
-    process_metrics: Option<&crate::pipeline::PipelineProcessMetrics>,
-    event: &Event,
-    ctx: &PipelineContext,
-    bump: &mut bumpalo::Bump,
-    dispatch_started_at: crate::time::UnixNanos,
-) -> Result<crate::pipeline::PipelineRunResult> {
-    // No `--test-pipeline` trace collector on the daemon hot path —
-    // passing `None` skips every trace push (and the `format!` /
-    // `to_string` work behind it) in `run_pipeline`, since nothing
-    // here reads `PipelineRunResult::trace`.
-    let mut result = match process_metrics {
-        Some(process_metrics) => crate::pipeline::run_pipeline_with_process_metrics_at(
-            pipeline,
-            event,
-            &ctx.config,
-            &ctx.funcs,
-            Some(&ctx.tap),
-            None,
-            crate::pipeline::OutputCapturePolicy::DiskOnly(&ctx.disk_outputs),
-            bump,
-            process_metrics,
-            dispatch_started_at,
-        )?,
-        None => crate::pipeline::run_pipeline_at(
-            pipeline,
-            event,
-            &ctx.config,
-            &ctx.funcs,
-            Some(&ctx.tap),
-            None,
-            crate::pipeline::OutputCapturePolicy::DiskOnly(&ctx.disk_outputs),
-            bump,
-            dispatch_started_at,
-        )?,
-    };
-
-    // Drain the per-event outputs vec into the queues. After this
-    // change every output statement enqueues a plain `OwnedEvent`
-    // regardless of the queue kind; render happens consumer-side
-    // inside each sink's `Output::consume`.
-    //
-    // `result.had_outputs` was set inside `run_pipeline` *before* the
-    // vec was moved out here, so the downstream
-    // `events_finished` / `events_discarded` decision still observes
-    // the original semantics (Finished AND emitted ≥1 output → finished).
-    let outputs = std::mem::take(&mut result.outputs);
-    // Each failed enqueue carries a per-output snapshot of just the
-    // fields the DLQ record actually stores (`OutputEvent`:
-    // `key`, `source`, `received_at`, `ingress`, `egress` — the same shape
-    // `inject output` replay needs). Snapshotting `OutputEvent` here
-    // rather than the whole `OwnedEvent` avoids the workspace
-    // `HashMap<String, OwnedValue>` deep clone on every success:
-    // pre-fix this ran unconditionally before `sender.send(event)`
-    // because the send consumed `event`, and populated workspaces
-    // dominated the runtime hot path (see the `634cbd0` perf
-    // regression thread). The input event the function received is
-    // not equivalent to a per-output snapshot: the pipeline may have
-    // produced a different `egress` per output, and replay must
-    // preserve the bytes the sink would have shipped — which is why
-    // the snapshot is per-`(output_name, event)`, not one shared
-    // capture at function entry.
-    let mut failed_outputs: Vec<(String, crate::pipeline::OutputEvent)> = Vec::new();
-    for (output_name, event) in outputs {
-        if let Some(sender) = ctx.output_senders.get(&output_name) {
-            // Snapshot the DLQ-relevant fields before the send
-            // consumes `event`. Cheap: five Copy scalars + two
-            // `Bytes` refcount bumps; workspace is not touched.
-            let snapshot = crate::pipeline::OutputEvent::from_owned(&event);
-            if let Err(e) = sender.send(event).await {
-                // QueueSender::send already bumped per-output
-                // `events_failed` on the Err branch — that gives the
-                // operator per-output visibility. Collect the names
-                // here so the pipeline-level disposition (below)
-                // routes the event through the DLQ instead of
-                // counting it as `events_finished`.
-                error!(
-                    "pipeline '{}': enqueue to output '{}' failed: {}",
-                    pipeline.name, output_name, e
-                );
-                failed_outputs.push((output_name, snapshot));
-            }
-        } else {
-            // Unknown output name slipped past startup validation —
-            // bug, but recoverable: treat as an enqueue failure so
-            // the event still hits the DLQ.
-            error!(
-                "pipeline '{}': output '{}' not found",
-                pipeline.name, output_name
-            );
-            failed_outputs.push((
-                output_name,
-                crate::pipeline::OutputEvent::from_owned(&event),
-            ));
-        }
-    }
-
-    // Any enqueue failure overrides the termination: the pipeline
-    // body finished, but the event never reached (some of) the
-    // configured downstream queues. Routing through the existing
-    // `Errored` termination path gives the operator the same
-    // recovery affordances as a `process` error — DLQ entry plus an
-    // `events_errored` increment — instead of a silent
-    // `events_finished` count on an event that was effectively lost.
-    //
-    // The DLQ records are emitted **per failed output** so each one
-    // can be replayed independently through
-    // `limpidctl inject output <name>` — joining them into one
-    // multi-output record would force the operator to re-run every
-    // sibling sink for one enqueue failure.
-    if !failed_outputs.is_empty() {
-        let reason = "output enqueue failed (queue closed, disk write \
-             error, or unknown output)"
-            .to_string();
-        result.termination = crate::pipeline::PipelineTermination::Errored;
-        for (output_name, snapshot) in failed_outputs {
-            result
-                .errored
-                .push(crate::pipeline::ErroredEventContext::Output {
-                    timestamp: chrono::Utc::now(),
-                    pipeline: pipeline.name.clone(),
-                    site: format!("{} enqueue", output_name),
-                    reason: reason.clone(),
-                    output_name: output_name.clone(),
-                    event: snapshot,
-                });
-        }
-    }
-
-    Ok(result)
-}
-
-async fn process_event(
-    event: &Event,
-    workers: &[Arc<PipelineWorker>],
-    ctx: &PipelineContext,
-    input_tap_key: &str,
-    input_queue_timer: &crate::metrics::InputQueueTimer,
-    bump: &mut bumpalo::Bump,
-) {
-    let dispatch_started_at = crate::time::UnixNanos::now();
-    process_event_at(
-        event,
-        workers,
-        ctx,
-        input_tap_key,
-        input_queue_timer,
-        bump,
-        dispatch_started_at,
-    )
-    .await;
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn process_event_at(
-    event: &Event,
-    workers: &[Arc<PipelineWorker>],
-    ctx: &PipelineContext,
-    input_tap_key: &str,
-    input_queue_timer: &crate::metrics::InputQueueTimer,
-    bump: &mut bumpalo::Bump,
-    dispatch_started_at: crate::time::UnixNanos,
-) {
-    input_queue_timer.observe_between(
-        crate::time::UnixNanos::from_datetime(event.received_at),
-        dispatch_started_at,
-    );
-    ctx.tap.emit(input_tap_key, event).await;
-    for (i, worker) in workers.iter().enumerate() {
-        worker.metrics.events_received.inc();
-
-        // Pass the event by reference. `run_pipeline` views it into
-        // the arena (read-only on the input owned form) and any DLQ
-        // path constructs a fresh `OwnedEvent` via `to_owned()` from
-        // the arena view — so multiple fan-out workers can share the
-        // same input without cloning the workspace HashMap.
-        //
-        // Reset only between fan-out workers (skip before the first):
-        // the caller's outer reset already cleared the bump for this
-        // event; the redundant inner reset on single-worker fan-out
-        // would just thrash the chunk-chain pointer for nothing.
-        if i > 0 {
-            bump.reset();
-        }
-        #[cfg(test)]
-        if let Some(gate) = &worker.serial_test_gate {
-            gate.wait().await;
-            gate.wait().await;
-        }
-        worker.metrics.inflight.inc();
-        match run_pipeline_with_outputs_inner(
-            &worker.def,
-            worker.process_metrics.as_deref(),
-            event,
-            ctx,
-            bump,
-            dispatch_started_at,
-        )
-        .await
-        {
-            Ok(result) => {
-                use crate::pipeline::PipelineTermination;
-                match result.termination {
-                    PipelineTermination::Dropped => {
-                        worker.metrics.events_dropped.inc();
-                    }
-                    PipelineTermination::Errored => {
-                        worker.metrics.events_errored.inc();
-                        // Drain every accumulated DLQ record. For a
-                        // pipeline-side failure this is exactly one
-                        // record; for a runtime-side per-failed-output
-                        // enqueue failure it is one record per output.
-                        // The metric increments once per pipeline run
-                        // (= one logical event lost), independent of
-                        // the record count, matching prior semantics.
-                        if result.errored.is_empty() {
-                            error!(
-                                "pipeline '{}': Errored termination without error context — bug",
-                                worker.def.name
-                            );
-                        } else {
-                            for err_ctx in &result.errored {
-                                write_errored_to_dlq(
-                                    err_ctx,
-                                    &worker.metrics,
-                                    ctx.error_log.as_ref(),
-                                    ctx.error_log_fallback,
-                                )
-                                .await;
-                            }
-                        }
-                    }
-                    PipelineTermination::Finished => {
-                        if !result.had_outputs {
-                            worker.metrics.events_discarded.inc();
-                        } else {
-                            worker.metrics.events_finished.inc();
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                // Pipeline body raised a runtime error that wasn't
-                // caught by `process` (= came out of expression
-                // evaluation in `error <expr>`, switch discriminant/
-                // pattern, or `if` condition).
-                // Pre-fix this branch only logged and the event
-                // disappeared without an `events_errored` increment
-                // or a DLQ entry — operators had no replay path,
-                // contradicting the documented runtime-error
-                // contract. Route through the same DLQ path as
-                // `PipelineTermination::Errored` so the metric +
-                // file record stay consistent across both shapes
-                // of runtime error.
-                worker.metrics.events_errored.inc();
-                let owned = event.to_owned();
-                let err_ctx = crate::pipeline::ErroredEventContext::Process {
-                    timestamp: chrono::Utc::now(),
-                    pipeline: worker.def.name.clone(),
-                    site: "(pipeline body)".to_string(),
-                    reason: e.to_string(),
-                    event: crate::pipeline::ProcessEvent::from_owned(&owned),
-                };
-                write_errored_to_dlq(
-                    &err_ctx,
-                    &worker.metrics,
-                    ctx.error_log.as_ref(),
-                    ctx.error_log_fallback,
-                )
-                .await;
-            }
-        }
-        worker.metrics.inflight.dec();
-    }
-}
-
-/// Persist an errored event to the dead-letter queue, or — if no
-/// `error_log` is configured — delegate to the `error_log_fallback`
-/// ladder helper so the failure surfaces as a payload-free operator
-/// signal by default (or as `Meta` / `Full` on explicit opt-in). The
-/// tracing line is best-effort, not a durable recovery trail; the
-/// DLQ file remains the load-bearing recovery target.
-///
-/// Shared by both runtime-error shapes the orchestrator surfaces:
-/// `PipelineTermination::Errored` (a `process` raised an error mid-
-/// pipeline) and an `Err` return from `run_pipeline` itself
-/// (expression evaluation under `error`/`if`/`switch` or a process
-/// body expression raised an error). Keeping the two on the same routing helper
-/// guarantees the operator-visible behaviour stays in lockstep:
-/// same JSONL on disk, same `events_errored` counter, same
-/// `events_errored_unwritable` semantics when the DLQ write itself
-/// fails.
-async fn write_errored_to_dlq(
-    err_ctx: &crate::pipeline::ErroredEventContext,
-    worker_metrics: &PipelineMetrics,
-    error_log: Option<&Arc<crate::error_log::ErrorLogWriter>>,
-    error_log_fallback: crate::error_log::ErrorLogFallback,
-) {
-    match error_log {
-        Some(writer) => {
-            if let Err(e) = writer.write(err_ctx).await {
-                worker_metrics.events_errored_unwritable.inc();
-                crate::modules::emit_dlq_tracing_fallback(
-                    /* error_log_configured */ true,
-                    error_log_fallback,
-                    err_ctx,
-                    None,
-                    Some(&e),
-                );
-            }
-        }
-        None => {
-            // No DLQ configured — payload-free tracing per ladder
-            // row-A. Operator declared no durable recovery is
-            // required; the summary line is all that surfaces.
-            crate::modules::emit_dlq_tracing_fallback(
-                /* error_log_configured */ false,
-                error_log_fallback,
-                err_ctx,
-                None,
-                None,
-            );
-        }
-    }
-}
-
-/// Return the list of input names a pipeline subscribes to (fan-in).
-///
-/// Empty if no `input` statement is present. A pipeline declared with
-/// `input a, b;` returns `["a", "b"]`; the legacy single-input form
-/// `input a;` returns `["a"]`.
-fn get_pipeline_inputs(pipeline: &PipelineDef) -> &[String] {
-    for stmt in &pipeline.body {
-        if let PipelineStatement::Input(names) = stmt {
-            return names;
-        }
-    }
-    &[]
-}
+mod startup;
+pub(crate) use startup::init_tables;
+
+mod pipeline_worker;
+use pipeline_worker::*;
 
 #[cfg(test)]
 mod tests {
@@ -2601,29 +1087,6 @@ def pipeline relay {{ input source; output to_b }}
         node_b.shutdown().await;
         delivered.expect("two-daemon LTP delivery timed out");
         assert_eq!(tokio::fs::read(&delivered_path).await.unwrap(), frame);
-    }
-
-    #[test]
-    fn ltp_peer_metric_registration_uses_the_deduplicated_static_config_union() {
-        use base64::Engine as _;
-
-        let mut spki = crate::ltp::ED25519_SPKI_PREFIX.to_vec();
-        spki.extend_from_slice(&[7; 32]);
-        let pubkey = base64::engine::general_purpose::STANDARD.encode(spki);
-        let config = compiled_config(&format!(
-            "def input ltp_in {{ type ltp peer {{ node_id \"peer-input\" pubkey {pubkey:?} }} }}\n\
-             def output ltp_out_a {{ type ltp peer {{ node_id \"peer-a\" pubkey {pubkey:?} endpoint \"127.0.0.1:7514\" }} }}\n\
-             def output ltp_out_b {{ type ltp peer {{ node_id \"peer-b\" pubkey {pubkey:?} endpoint \"127.0.0.1:7515\" }} }}"
-        ));
-
-        assert_eq!(
-            configured_ltp_peer_ids(&config).unwrap(),
-            BTreeSet::from([
-                "peer-a".to_owned(),
-                "peer-b".to_owned(),
-                "peer-input".to_owned(),
-            ])
-        );
     }
 
     fn metric_series(registry: &Registry, name: &str) -> Vec<serde_json::Value> {
@@ -4070,29 +2533,6 @@ def pipeline p {{ input source; output sink }}
         assert_eq!(metric_series(&registry, "limpid_build_info").len(), 1);
     }
 
-    #[test]
-    fn get_pipeline_inputs_single() {
-        let def = pipeline_def("def pipeline p { input a; drop }");
-        assert_eq!(get_pipeline_inputs(&def), &["a".to_string()]);
-    }
-
-    #[test]
-    fn get_pipeline_inputs_fan_in() {
-        let def = pipeline_def("def pipeline p { input a, b, c; drop }");
-        assert_eq!(
-            get_pipeline_inputs(&def),
-            &["a".to_string(), "b".to_string(), "c".to_string()]
-        );
-    }
-
-    #[test]
-    fn get_pipeline_inputs_missing_input_stmt() {
-        // Not valid per compiled-config validation, but the helper alone should
-        // still return an empty slice rather than panic.
-        let def = pipeline_def("def pipeline p { drop }");
-        assert!(get_pipeline_inputs(&def).is_empty());
-    }
-
     /// End-to-end-ish fan-in runtime test: two independent mpsc channels
     /// (simulating two input sources) both push events into a dispatcher
     /// that shares a single `PipelineWorker`. Events from both sides land
@@ -4481,7 +2921,7 @@ def pipeline p {{ input source; output sink }}
     /// `queue::tests::tokio_mpsc_close_then_permit_send_still_visible`.
     #[test]
     fn pipeline_worker_shutdown_arm_uses_close_recv_pattern() {
-        let src = include_str!("runtime.rs");
+        let src = include_str!("runtime/pipeline_worker.rs");
         // Anchor on the specific inner select arm inside
         // `run_pipeline_workers`, not the outer `let event = tokio::select!`
         // that races receive vs shutdown.
@@ -4647,7 +3087,11 @@ def pipeline second { input i; output sink; finish }
 
     #[test]
     fn startup_is_guarded_until_all_listener_handles_are_registered() {
-        let src = include_str!("runtime.rs");
+        let src = format!(
+            "{}{}",
+            include_str!("runtime/lifecycle.rs"),
+            include_str!("runtime/startup.rs")
+        );
         assert!(src.contains(&["struct Startup", "Guard"].concat()));
         assert!(src.contains(&["startup_guard.", "commit"].concat()));
         assert!(src.contains(&["late_post_listener", "_failure"].concat()));
@@ -4655,8 +3099,12 @@ def pipeline second { input i; output sink; finish }
 
     #[test]
     fn startup_transaction_mutant_sensitivity() {
-        let src = include_str!("runtime.rs");
-        let production = &src[..src.find("#[cfg(test)]\nmod tests").unwrap()];
+        let lifecycle = include_str!("runtime/lifecycle.rs");
+        let startup = include_str!("runtime/startup.rs");
+        let lifecycle = &lifecycle[..lifecycle.find("#[cfg(test)]\nmod tests").unwrap()];
+        let startup = &startup[..startup.find("#[cfg(test)]\nmod tests").unwrap()];
+        let production = format!("{lifecycle}{startup}");
+        let production = production.as_str();
         let guard = production.find("StartupGuard::new").unwrap();
         assert!(
             production.find("validate_control_socket_parent").unwrap() < guard,
@@ -4716,7 +3164,7 @@ def pipeline second { input i; output sink; finish }
             .expect("closed watch must resolve without spinning")
         );
         let marker = ["shutdown_change_is_terminal", "(&mut shutdown)"].concat();
-        assert!(include_str!("runtime.rs").contains(&marker));
+        assert!(include_str!("runtime/pipeline_worker.rs").contains(&marker));
     }
 
     #[tokio::test]
